% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R, R/reexports_bricklayer.R,
%   R/reexports_rmoriedata.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{irtsp}
\alias{morie_arsau_analyze_aggregate_summary}
\alias{morie_arsau_analyze_detailed_dataset}
\alias{morie_arsau_analyze_individual_records}
\alias{morie_arsau_analyze_main_records}
\alias{morie_arsau_analyze_probe_cycle_records}
\alias{morie_arsau_analyze_weapon_records}
\alias{morie_arsau_available_datasets}
\alias{morie_arsau_available_years}
\alias{morie_arsau_ckan_url}
\alias{morie_arsau_describe}
\alias{morie_arsau_download}
\alias{morie_arsau_fetch_sidecar}
\alias{morie_arsau_load_aggregate_summary}
\alias{morie_arsau_load_detailed_dataset}
\alias{morie_arsau_load_individual_records}
\alias{morie_arsau_load_main_records}
\alias{morie_arsau_load_probe_cycle_records}
\alias{morie_arsau_load_weapon_records}
\alias{morie_arsau_read_markdown_dictionary}
\alias{morie_arsau_read_sidecar}
\alias{morie_arsau_read_xlsx_dictionary}
\alias{morie_arsau_registry_df}
\alias{morie_arsau_sidecar_schema}
\alias{morie_arsau_sidecar_to_frame}
\alias{morie_calculate_ebac}
\alias{morie_ckan_search}
\alias{morie_cpads_canonicalize_frame}
\alias{morie_cpads_contract}
\alias{morie_cpads_has_raw_columns}
\alias{morie_cpads_infer_file_format}
\alias{morie_cpads_missing_variables}
\alias{morie_cpads_validate_frame}
\alias{morie_dataset_catalog}
\alias{morie_dataset_column_profile}
\alias{morie_dataset_detect_role}
\alias{morie_dataset_infer_level}
\alias{morie_dataset_load}
\alias{morie_dataset_portal_catalog}
\alias{morie_dataset_portal_catalog_clear_cache}
\alias{morie_dataset_profile}
\alias{morie_dataset_profile_summary_table}
\alias{morie_dataset_profile_to_list}
\alias{morie_dataset_suggest_plan}
\alias{morie_dataset_summarize_column}
\alias{morie_datasets_arcgis_item_by_id}
\alias{morie_datasets_arcgis_item_metadata}
\alias{morie_datasets_arsau_aggregate_summary}
\alias{morie_datasets_arsau_detailed_dataset}
\alias{morie_datasets_arsau_uof_individual_records}
\alias{morie_datasets_arsau_uof_main_records}
\alias{morie_datasets_arsau_uof_probe_cycle_records}
\alias{morie_datasets_arsau_uof_weapon_records}
\alias{morie_datasets_bigquery}
\alias{morie_datasets_browse}
\alias{morie_datasets_calgary_community_crime_stats}
\alias{morie_datasets_calgary_fire_response_calls}
\alias{morie_datasets_calgary_fire_stations}
\alias{morie_datasets_calgary_open_crime_adjacent_layers}
\alias{morie_datasets_calgary_opendata_bulk_layers}
\alias{morie_datasets_calgary_socrata_by_id}
\alias{morie_datasets_chicago_arrests}
\alias{morie_datasets_chicago_community_areas}
\alias{morie_datasets_chicago_crime}
\alias{morie_datasets_chicago_crime_map}
\alias{morie_datasets_chicago_crime_odata}
\alias{morie_datasets_chicago_crime_resolved}
\alias{morie_datasets_chicago_crime_soql}
\alias{morie_datasets_chicago_iucr_codes}
\alias{morie_datasets_chicago_neighborhoods}
\alias{morie_datasets_chicago_opendata_bulk_layers}
\alias{morie_datasets_chicago_police_beats}
\alias{morie_datasets_chicago_police_districts}
\alias{morie_datasets_chicago_socrata_by_id}
\alias{morie_datasets_chicago_wards}
\alias{morie_datasets_ckan_package}
\alias{morie_datasets_ckan_search}
\alias{morie_datasets_cpads}
\alias{morie_datasets_cpd_public_arrests}
\alias{morie_datasets_edmonton_fire_stations}
\alias{morie_datasets_edmonton_open_crime_adjacent_layers}
\alias{morie_datasets_edmonton_opendata_bulk_layers}
\alias{morie_datasets_edmonton_police_stations}
\alias{morie_datasets_edmonton_socrata_by_id}
\alias{morie_datasets_external_socrata_layers}
\alias{morie_datasets_load_by_key}
\alias{morie_datasets_montreal_ckan_resource}
\alias{morie_datasets_montreal_justice_safety_layers}
\alias{morie_datasets_montreal_opendata_bulk_layers}
\alias{morie_datasets_montreal_sim_intervention_types}
\alias{morie_datasets_montreal_sim_interventions}
\alias{morie_datasets_namus_missing_persons}
\alias{morie_datasets_nibrs}
\alias{morie_datasets_nist_rds}
\alias{morie_datasets_nyc_boroughs}
\alias{morie_datasets_nyc_boundaries_catalog}
\alias{morie_datasets_nyc_community_districts}
\alias{morie_datasets_nyc_council_districts}
\alias{morie_datasets_nyc_ntas_2020}
\alias{morie_datasets_nyc_nypd_arrests_historic}
\alias{morie_datasets_nyc_nypd_arrests_ytd}
\alias{morie_datasets_nyc_nypd_boro_crosswalk}
\alias{morie_datasets_nyc_nypd_by_key}
\alias{morie_datasets_nyc_nypd_complaint_historic}
\alias{morie_datasets_nyc_nypd_complaint_ytd}
\alias{morie_datasets_nyc_nypd_hate_crimes}
\alias{morie_datasets_nyc_nypd_law_books}
\alias{morie_datasets_nyc_nypd_layers}
\alias{morie_datasets_nyc_nypd_offense_codes}
\alias{morie_datasets_nyc_nypd_resolved}
\alias{morie_datasets_nyc_nypd_uof_incidents}
\alias{morie_datasets_nyc_nypd_uof_subjects}
\alias{morie_datasets_nyc_nypd_vehicle_stops}
\alias{morie_datasets_nyc_opendata_bulk_layers}
\alias{morie_datasets_nyc_police_precincts}
\alias{morie_datasets_nyc_school_districts}
\alias{morie_datasets_nyc_socrata_by_id}
\alias{morie_datasets_nyc_stop_and_frisk}
\alias{morie_datasets_nyc_zctas}
\alias{morie_datasets_ontario_ckan_by_key}
\alias{morie_datasets_ontario_ckan_layers}
\alias{morie_datasets_otis_a01}
\alias{morie_datasets_otis_a01_restrictive_confinement}
\alias{morie_datasets_otis_b01_segregation_detailed}
\alias{morie_datasets_otis_b02_segregation_total_days}
\alias{morie_datasets_otis_b03_seg_alerts_by_institution}
\alias{morie_datasets_otis_b04_seg_consecutive_by_region}
\alias{morie_datasets_otis_b05_seg_consecutive_lengths}
\alias{morie_datasets_otis_b06_seg_reason_by_institution}
\alias{morie_datasets_otis_b07_seg_alerts_by_gender}
\alias{morie_datasets_otis_b08_seg_consecutive_by_institution}
\alias{morie_datasets_otis_b09_seg_n_times}
\alias{morie_datasets_otis_c01_individuals_total}
\alias{morie_datasets_otis_c02_individuals_by_institution}
\alias{morie_datasets_otis_c03_individuals_race_by_gender}
\alias{morie_datasets_otis_c04_individuals_race_by_region}
\alias{morie_datasets_otis_c05_individuals_religion_by_region}
\alias{morie_datasets_otis_c06_individuals_age_by_region}
\alias{morie_datasets_otis_c07_individuals_alerts}
\alias{morie_datasets_otis_c08_individuals_religion_by_gender}
\alias{morie_datasets_otis_c09_individuals_age_by_gender}
\alias{morie_datasets_otis_c10_aggregate_durations_by_institution}
\alias{morie_datasets_otis_c11_aggregate_lengths}
\alias{morie_datasets_otis_c12_aggregate_durations_by_region}
\alias{morie_datasets_otis_d01_deaths_in_custody}
\alias{morie_datasets_otis_d02_deaths_by_gender}
\alias{morie_datasets_otis_d03_deaths_by_race}
\alias{morie_datasets_otis_d04_deaths_by_religion}
\alias{morie_datasets_otis_d05_deaths_by_age_category}
\alias{morie_datasets_otis_d06_cause_by_alert}
\alias{morie_datasets_otis_d07_alerts_by_housing_unit}
\alias{morie_datasets_ottawa_open_crime_adjacent_layers}
\alias{morie_datasets_ottawa_opendata_bulk_layers}
\alias{morie_datasets_siu_director_reports}
\alias{morie_datasets_siu_report_fields}
\alias{morie_datasets_siu_report_text}
\alias{morie_datasets_statcan_ccjs_cubes}
\alias{morie_datasets_statcan_cube_metadata}
\alias{morie_datasets_statcan_full_csv_url}
\alias{morie_datasets_statcan_vectors}
\alias{morie_datasets_summary}
\alias{morie_datasets_toronto_ambulance_stations}
\alias{morie_datasets_toronto_asr_miscellaneous}
\alias{morie_datasets_toronto_open_ckan_resource}
\alias{morie_datasets_toronto_open_crime_adjacent_layers}
\alias{morie_datasets_toronto_opendata_bulk_layers}
\alias{morie_datasets_toronto_zoning_per_neighbourhood}
\alias{morie_datasets_tps_2008_firs}
\alias{morie_datasets_tps_2009_firs}
\alias{morie_datasets_tps_2010_firs}
\alias{morie_datasets_tps_2011_firs}
\alias{morie_datasets_tps_2012_firs}
\alias{morie_datasets_tps_2013_firs}
\alias{morie_datasets_tps_administrative}
\alias{morie_datasets_tps_arcgis_hub_by_id}
\alias{morie_datasets_tps_arcgis_hub_download}
\alias{morie_datasets_tps_arcgis_hub_layers}
\alias{morie_datasets_tps_arrested_and_charged_persons}
\alias{morie_datasets_tps_arrests_and_strip_searches}
\alias{morie_datasets_tps_assault}
\alias{morie_datasets_tps_automobile_ksi}
\alias{morie_datasets_tps_autotheft}
\alias{morie_datasets_tps_bicycle_thefts}
\alias{morie_datasets_tps_bicycletheft}
\alias{morie_datasets_tps_breakandenter}
\alias{morie_datasets_tps_budget_2020}
\alias{morie_datasets_tps_budget_2021}
\alias{morie_datasets_tps_budget_2022}
\alias{morie_datasets_tps_budget_2023}
\alias{morie_datasets_tps_budget_2024}
\alias{morie_datasets_tps_budget_2025}
\alias{morie_datasets_tps_budget_2026}
\alias{morie_datasets_tps_budget_by_command}
\alias{morie_datasets_tps_calls_for_service_attended}
\alias{morie_datasets_tps_community_safety_indicators}
\alias{morie_datasets_tps_complaint_dispositions}
\alias{morie_datasets_tps_cyclist_ksi}
\alias{morie_datasets_tps_dispatched_calls_by_division}
\alias{morie_datasets_tps_facilities}
\alias{morie_datasets_tps_fatals_ksi}
\alias{morie_datasets_tps_firearms_top_calibres}
\alias{morie_datasets_tps_gross_expenditures_by_division}
\alias{morie_datasets_tps_gross_operating_budget}
\alias{morie_datasets_tps_hatecrimes}
\alias{morie_datasets_tps_homicide}
\alias{morie_datasets_tps_homicides}
\alias{morie_datasets_tps_intimate_partner_family_violence}
\alias{morie_datasets_tps_investigated_alleged_misconduct}
\alias{morie_datasets_tps_killed_and_seriously_injured}
\alias{morie_datasets_tps_layers}
\alias{morie_datasets_tps_major_crime}
\alias{morie_datasets_tps_mha_apprehensions}
\alias{morie_datasets_tps_miscellaneous_calls_for_service}
\alias{morie_datasets_tps_miscellaneous_data}
\alias{morie_datasets_tps_miscellaneous_firearms}
\alias{morie_datasets_tps_motorcylist_ksi}
\alias{morie_datasets_tps_neighbourhood_crime_rates}
\alias{morie_datasets_tps_passenger_ksi}
\alias{morie_datasets_tps_patrol_zone}
\alias{morie_datasets_tps_pedestrian_ksi}
\alias{morie_datasets_tps_personnel_by_command}
\alias{morie_datasets_tps_personnel_by_rank}
\alias{morie_datasets_tps_personnel_by_rank_by_division}
\alias{morie_datasets_tps_persons_in_crisis_calls_for_service_attended}
\alias{morie_datasets_tps_police_divisions}
\alias{morie_datasets_tps_psdp_resolved}
\alias{morie_datasets_tps_regulated_interactions}
\alias{morie_datasets_tps_reported_crimes}
\alias{morie_datasets_tps_robbery}
\alias{morie_datasets_tps_search_of_persons}
\alias{morie_datasets_tps_shooting_firearm_discharges}
\alias{morie_datasets_tps_shootings}
\alias{morie_datasets_tps_staffing_by_command}
\alias{morie_datasets_tps_theft_from_motor_vehicle}
\alias{morie_datasets_tps_theft_over}
\alias{morie_datasets_tps_tickets_issued}
\alias{morie_datasets_tps_top_20_offences_of_firearm_seizures}
\alias{morie_datasets_tps_total_public_complaints}
\alias{morie_datasets_tps_traffic_collisions}
\alias{morie_datasets_tps_use_of_force_call_for_service_types}
\alias{morie_datasets_tps_use_of_force_call_sources_by_month}
\alias{morie_datasets_tps_use_of_force_gender_composition}
\alias{morie_datasets_tps_use_of_force_location_of_occurrences}
\alias{morie_datasets_tps_use_of_force_occurrence_category}
\alias{morie_datasets_tps_use_of_force_time_of_day_trends}
\alias{morie_datasets_tps_use_of_force_use_of_force_types_and_perceived_weapons}
\alias{morie_datasets_tps_victims_of_crime}
\alias{morie_datasets_vancouver_community_centres}
\alias{morie_datasets_vancouver_community_food_markets}
\alias{morie_datasets_vancouver_disability_parking}
\alias{morie_datasets_vancouver_fire_halls}
\alias{morie_datasets_vancouver_graffiti}
\alias{morie_datasets_vancouver_homeless_shelters}
\alias{morie_datasets_vancouver_noise_control_areas}
\alias{morie_datasets_vancouver_opendata_bulk_layers}
\alias{morie_datasets_vancouver_opendata_by_id}
\alias{morie_datasets_vancouver_opendata_layers}
\alias{morie_datasets_vancouver_property_use_inspection_districts}
\alias{morie_datasets_vancouver_public_art}
\alias{morie_datasets_vpd_crime}
\alias{morie_datasets_vpd_legal_disclaimer}
\alias{morie_fairness_apply_profile}
\alias{morie_fairness_average_odds_difference}
\alias{morie_fairness_bias_amplification}
\alias{morie_fairness_city_profile}
\alias{morie_fairness_column_map}
\alias{morie_fairness_demographic_parity}
\alias{morie_fairness_disparate_impact}
\alias{morie_fairness_equalized_odds}
\alias{morie_fairness_get_city}
\alias{morie_fairness_gini}
\alias{morie_fairness_list_cities}
\alias{morie_fairness_noisy_or_detection}
\alias{morie_fairness_predpol_aggregate_areas}
\alias{morie_fairness_predpol_calibration_audit}
\alias{morie_fairness_predpol_score_disparity}
\alias{morie_fairness_predpol_temporal_audit}
\alias{morie_fairness_register_city}
\alias{morie_fairness_simulate_biased_crime_data}
\alias{morie_fairness_xai_ale}
\alias{morie_fairness_xai_ceteris_paribus}
\alias{morie_fairness_xai_partial_dependence}
\alias{morie_fairness_xai_permutation_importance}
\alias{morie_fairness_xai_shap_values}
\alias{morie_fetch}
\alias{morie_fetch_arcgis}
\alias{morie_fetch_siu}
\alias{morie_infer_measurement_level}
\alias{morie_ingest_bigquery_build_sql}
\alias{morie_ingest_bigquery_query}
\alias{morie_ingest_bigquery_table}
\alias{morie_ingest_chicago_crime}
\alias{morie_ingest_chicago_crime_bigquery}
\alias{morie_ingest_chicago_resources}
\alias{morie_ingest_chicago_socrata}
\alias{morie_ingest_cihi_xlsx}
\alias{morie_ingest_ckan_fetch_package_csvs}
\alias{morie_ingest_ckan_package_search}
\alias{morie_ingest_ckan_package_show}
\alias{morie_ingest_ckan_read_resource}
\alias{morie_ingest_ckan_resource_show}
\alias{morie_ingest_ckan_search_packages}
\alias{morie_ingest_forensics_namus_missing}
\alias{morie_ingest_forensics_nibrs}
\alias{morie_ingest_forensics_nist_rds}
\alias{morie_ingest_statcan_cansim}
\alias{morie_ingest_statcan_csv}
\alias{morie_ingest_tps_feature_layer}
\alias{morie_ingest_tps_fetch}
\alias{morie_ingest_tps_layers}
\alias{morie_is_over_legal_limit}
\alias{morie_laniyonu_actuarial_risk_disparity}
\alias{morie_laniyonu_gentrification_policing}
\alias{morie_laniyonu_smi_force_disparity}
\alias{morie_otis_aipw_ate}
\alias{morie_otis_aipw_superlearner}
\alias{morie_otis_alert_state_combo}
\alias{morie_otis_all_analyses}
\alias{morie_otis_analyze_a01}
\alias{morie_otis_analyze_a01_ruhela_alt_age}
\alias{morie_otis_analyze_a01_ruhela_alt_gender}
\alias{morie_otis_analyze_a01_ruhela_alt_toronto}
\alias{morie_otis_analyze_a01_ruhela_formulations}
\alias{morie_otis_analyze_a01_ruhela_per_year}
\alias{morie_otis_analyze_a01_ruhela_subgroup_female}
\alias{morie_otis_analyze_a01_ruhela_subgroup_male}
\alias{morie_otis_analyze_a01_with_csi_context}
\alias{morie_otis_analyze_all}
\alias{morie_otis_analyze_b01}
\alias{morie_otis_analyze_b01_ruhela_alt_age}
\alias{morie_otis_analyze_b01_ruhela_alt_gender}
\alias{morie_otis_analyze_b01_ruhela_alt_toronto}
\alias{morie_otis_analyze_b01_ruhela_formulations}
\alias{morie_otis_analyze_b01_ruhela_per_year}
\alias{morie_otis_analyze_b01_ruhela_subgroup_female}
\alias{morie_otis_analyze_b01_ruhela_subgroup_male}
\alias{morie_otis_analyze_b02}
\alias{morie_otis_analyze_b02_ruhela_alt_age}
\alias{morie_otis_analyze_b02_ruhela_alt_region}
\alias{morie_otis_analyze_b02_ruhela_formulations}
\alias{morie_otis_analyze_b03}
\alias{morie_otis_analyze_b03_ruhela_aggregate}
\alias{morie_otis_analyze_b04}
\alias{morie_otis_analyze_b04_ruhela_aggregate}
\alias{morie_otis_analyze_b05}
\alias{morie_otis_analyze_b05_mandela_classification}
\alias{morie_otis_analyze_b05_ruhela_aggregate}
\alias{morie_otis_analyze_b06}
\alias{morie_otis_analyze_b06_ruhela_aggregate}
\alias{morie_otis_analyze_b07}
\alias{morie_otis_analyze_b07_ruhela_aggregate}
\alias{morie_otis_analyze_b08}
\alias{morie_otis_analyze_b08_ruhela_aggregate}
\alias{morie_otis_analyze_b09}
\alias{morie_otis_analyze_b09_ruhela_aggregate}
\alias{morie_otis_analyze_c_chi2}
\alias{morie_otis_analyze_c01}
\alias{morie_otis_analyze_c01_ruhela_aggregate}
\alias{morie_otis_analyze_c01_ruhela_aggregate_region_cluster}
\alias{morie_otis_analyze_c02}
\alias{morie_otis_analyze_c02_ruhela_aggregate}
\alias{morie_otis_analyze_c03}
\alias{morie_otis_analyze_c03_ruhela_aggregate}
\alias{morie_otis_analyze_c04_ruhela_aggregate}
\alias{morie_otis_analyze_c04_ruhela_aggregate_region_cluster}
\alias{morie_otis_analyze_c05_ruhela_aggregate}
\alias{morie_otis_analyze_c06_ruhela_aggregate}
\alias{morie_otis_analyze_c07}
\alias{morie_otis_analyze_c07_ruhela_aggregate}
\alias{morie_otis_analyze_c08_ruhela_aggregate}
\alias{morie_otis_analyze_c09_ruhela_aggregate}
\alias{morie_otis_analyze_c10}
\alias{morie_otis_analyze_c10_ruhela_aggregate}
\alias{morie_otis_analyze_c11}
\alias{morie_otis_analyze_c11_mandela_classification}
\alias{morie_otis_analyze_c11_ruhela_aggregate}
\alias{morie_otis_analyze_c12}
\alias{morie_otis_analyze_c12_ruhela_aggregate}
\alias{morie_otis_analyze_d_chi2}
\alias{morie_otis_analyze_d01}
\alias{morie_otis_analyze_d02_ruhela_aggregate}
\alias{morie_otis_analyze_d03_ruhela_aggregate}
\alias{morie_otis_analyze_d04_ruhela_aggregate}
\alias{morie_otis_analyze_d05_ruhela_aggregate}
\alias{morie_otis_analyze_d06}
\alias{morie_otis_analyze_d07}
\alias{morie_otis_analyze_otis_mandela_provincial_vs_federal}
\alias{morie_otis_analyze_ruhela_grid}
\alias{morie_otis_analyze_ruhela_master}
\alias{morie_otis_analyze_ruhela_per_year}
\alias{morie_otis_analyzers}
\alias{morie_otis_astcmb}
\alias{morie_otis_causal_grid}
\alias{morie_otis_churn_analyze_all}
\alias{morie_otis_classify_mandela_combo}
\alias{morie_otis_descriptives}
\alias{morie_otis_disciplinary_medical_overlap}
\alias{morie_otis_dml}
\alias{morie_otis_embedding_distribution}
\alias{morie_otis_intra_year_transition_matrix}
\alias{morie_otis_ipw_ate}
\alias{morie_otis_irm_dml}
\alias{morie_otis_irr_glmm_vm}
\alias{morie_otis_load}
\alias{morie_otis_make_pair_a}
\alias{morie_otis_make_pair_alert_to_volatility_a01}
\alias{morie_otis_make_pair_alert_to_volatility_all}
\alias{morie_otis_make_pair_alert_to_volatility_naive}
\alias{morie_otis_make_pair_alert_to_volatility_ruhela}
\alias{morie_otis_make_pair_b}
\alias{morie_otis_make_pair_c}
\alias{morie_otis_mortification_cooccurrence}
\alias{morie_otis_otdesc}
\alias{morie_otis_otdml}
\alias{morie_otis_path_complexity_gini}
\alias{morie_otis_plr}
\alias{morie_otis_psm}
\alias{morie_otis_psm_subclass}
\alias{morie_otis_rc_trends}
\alias{morie_otis_rctrnd}
\alias{morie_otis_regC_demog_contingency}
\alias{morie_otis_region_alert_state_richness}
\alias{morie_otis_regional_placement}
\alias{morie_otis_repeat_placement_concentration}
\alias{morie_otis_rplace}
\alias{morie_otis_tps_analyze_all}
\alias{morie_otis_tps_composite_overlay}
\alias{morie_otis_tps_per_region_rollup}
\alias{morie_otis_tps_yoy_correlation}
\alias{morie_otis_volat}
\alias{morie_otis_volatility}
\alias{morie_otis_within_year_placement_count}
\alias{morie_otis_within_year_region_diversity}
\alias{morie_parse_nypd_law_code}
\alias{morie_predpol_aggregate_areas}
\alias{morie_predpol_calibration_audit}
\alias{morie_predpol_score_disparity}
\alias{morie_profile_dataset}
\alias{morie_siu_all_analyses}
\alias{morie_siu_anomaly_check}
\alias{morie_siu_audit_case}
\alias{morie_siu_audit_columns}
\alias{morie_siu_by_police_service}
\alias{morie_siu_by_year}
\alias{morie_siu_cache_path}
\alias{morie_siu_case_counts}
\alias{morie_siu_charges_by_year_chi2}
\alias{morie_siu_compare}
\alias{morie_siu_decision_timing}
\alias{morie_siu_demographics}
\alias{morie_siu_fetch_cases}
\alias{morie_siu_fetch_dataframe}
\alias{morie_siu_index}
\alias{morie_siu_index_url}
\alias{morie_siu_llm_extract}
\alias{morie_siu_mental_health_race_indicators}
\alias{morie_siu_parse_html}
\alias{morie_siu_parse_news_html}
\alias{morie_siu_record_correction}
\alias{morie_siu_refresh_manifest}
\alias{morie_siu_sanity_check}
\alias{morie_siu_translate}
\alias{morie_siu_translate_fr_to_en}
\alias{morie_siuiap_cite}
\alias{morie_siuiap_panel_summary}
\alias{morie_spatial_voting_aldrich_mckelvey}
\alias{morie_spatial_voting_alpha_nominate}
\alias{morie_spatial_voting_anchoring_vignettes}
\alias{morie_spatial_voting_bayesian_am}
\alias{morie_spatial_voting_bayesian_irt_likelihood}
\alias{morie_spatial_voting_bayesian_irt_posterior}
\alias{morie_spatial_voting_bayesian_mds}
\alias{morie_spatial_voting_bayesian_unfolding}
\alias{morie_spatial_voting_blackbox}
\alias{morie_spatial_voting_cjr_irt}
\alias{morie_spatial_voting_classical_mds}
\alias{morie_spatial_voting_cutting_lines}
\alias{morie_spatial_voting_double_centering}
\alias{morie_spatial_voting_dw_nominate}
\alias{morie_spatial_voting_dynamic_irt}
\alias{morie_spatial_voting_em_irt}
\alias{morie_spatial_voting_ideal_point_recovery}
\alias{morie_spatial_voting_indscal}
\alias{morie_spatial_voting_mds_fit_stats}
\alias{morie_spatial_voting_mlsmu6}
\alias{morie_spatial_voting_nominate_bootstrap}
\alias{morie_spatial_voting_nominate_loglik}
\alias{morie_spatial_voting_nominate_utility}
\alias{morie_spatial_voting_nominate_vote_prob}
\alias{morie_spatial_voting_nonmetric_mds}
\alias{morie_spatial_voting_nonparametric_bootstrap}
\alias{morie_spatial_voting_normal_vectors}
\alias{morie_spatial_voting_optimal_classification}
\alias{morie_spatial_voting_ordered_oc}
\alias{morie_spatial_voting_ordinal_irt}
\alias{morie_spatial_voting_procrustes}
\alias{morie_spatial_voting_smacof}
\alias{morie_spatial_voting_smacof_unfolding}
\alias{morie_spatial_voting_unfolding_stress}
\alias{morie_spatial_voting_wordfish}
\alias{morie_suggest_analysis_plan}
\alias{morie_synth_otis}
\alias{morie_synth_otis_all}
\alias{morie_to_hood_crosswalk}
\alias{morie_to_neighbourhoods}
\alias{morie_tps_add_hood_140_from_158}
\alias{morie_tps_add_hood_158_from_140}
\alias{morie_tps_aggregate_158_to_140}
\alias{morie_tps_analyze_all}
\alias{morie_tps_analyze_csi_from_dataframes}
\alias{morie_tps_analyze_one}
\alias{morie_tps_arima_forecast}
\alias{morie_tps_assert_hood_version}
\alias{morie_tps_available_formats}
\alias{morie_tps_bivariate_moran}
\alias{morie_tps_bivariate_morans_i}
\alias{morie_tps_category_correlation_matrix}
\alias{morie_tps_changepoint_detection}
\alias{morie_tps_compare_hawkes_kernels}
\alias{morie_tps_composite_index}
\alias{morie_tps_crime_compare}
\alias{morie_tps_criminal_network_graph}
\alias{morie_tps_csi_per_neighbourhood}
\alias{morie_tps_csi_per_year}
\alias{morie_tps_csi_weight}
\alias{morie_tps_data_dir}
\alias{morie_tps_dbscan_clusters}
\alias{morie_tps_disaggregate_140_to_158}
\alias{morie_tps_district_for_centroid}
\alias{morie_tps_fetch_category}
\alias{morie_tps_fetch_dataframe}
\alias{morie_tps_fokker_planck_grid}
\alias{morie_tps_getis_ord_g_star}
\alias{morie_tps_gini_concentration}
\alias{morie_tps_hawkes_advanced_fit}
\alias{morie_tps_hawkes_markovian_vs_nonmarkovian}
\alias{morie_tps_hawkes_temporal_fit}
\alias{morie_tps_inspection_game_phase}
\alias{morie_tps_kde_density}
\alias{morie_tps_langevin_simulate}
\alias{morie_tps_levy_flight_alpha}
\alias{morie_tps_list_categories}
\alias{morie_tps_list_datasets}
\alias{morie_tps_list_formats}
\alias{morie_tps_load}
\alias{morie_tps_load_dataset}
\alias{morie_tps_local_morans_i}
\alias{morie_tps_lotka_volterra_police_crime}
\alias{morie_tps_moran_sweep_heatmap}
\alias{morie_tps_morans_i_neighbourhood}
\alias{morie_tps_neighbourhood_concentration}
\alias{morie_tps_offence_summary}
\alias{morie_tps_polygon_morans_i}
\alias{morie_tps_pretty_label}
\alias{morie_tps_project_xy}
\alias{morie_tps_psdp_layers}
\alias{morie_tps_render_choropleth}
\alias{morie_tps_render_dbscan}
\alias{morie_tps_render_district_proportional}
\alias{morie_tps_render_points}
\alias{morie_tps_render_quad}
\alias{morie_tps_render_satscan_panel}
\alias{morie_tps_render_yearly_grid}
\alias{morie_tps_resolve_hood_col}
\alias{morie_tps_ripley_k}
\alias{morie_tps_sarima_forecast}
\alias{morie_tps_sdb_reaction_diffusion}
\alias{morie_tps_sdb_turing_demo}
\alias{morie_tps_seasonal_pattern}
\alias{morie_tps_spatial_summary}
\alias{morie_tps_statphysics_analyze_all}
\alias{morie_tps_temporal_summary}
\alias{morie_tps_urban_scaling_beta}
\alias{morie_tps_use_of_force}
\alias{morie_tps_year_over_year_trend}
\alias{morie_tps_year_to_hood_version}
\alias{morie_tps_yoy_panel}
\alias{morie_vpd_download_instructions}
\alias{mrm_classify_mandela}
\alias{mrm_otis_mandela_spectrum}
\alias{mrm_otis_mortification_cooccurrence}
\alias{mrm_otis_placement_concentration}
\alias{mrm_otis_region_locality}
\alias{mrm_otis_seg_duration_km}
\alias{mrm_siu_case_to_decision_km}
\alias{mrm_siu_outcome_classifier}
\alias{mrm_siu_per_service_rate}
\alias{mrm_tps_levy_scaling}
\alias{mrm_tps_load_hawkes_refit}
\alias{mrm_tps_moran_clustering}
\alias{mrm_tps_neighbourhood_recurrence_km}
\alias{tpspn}
\alias{apply_schema_validation}
\alias{download_data}
\alias{friendly_download}
\alias{load_provenance}
\alias{make_manifest}
\alias{make_synthetic_column}
\alias{make_synthetic_csv}
\alias{record}
\alias{resolve_via_ckan}
\alias{resolve_via_ckan_search}
\alias{sha256_file}
\alias{validate_schema}
\alias{verify_sha256}
\alias{write_manifest_json}
\alias{write_summary_txt}
\alias{morie_data_checksums}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{rmorie}{\code{\link[rmorie:irtsp]{irtsp()}}, \code{\link[rmorie:morie_arsau_analyze_aggregate_summary]{morie_arsau_analyze_aggregate_summary()}}, \code{\link[rmorie:morie_arsau_analyze_detailed_dataset]{morie_arsau_analyze_detailed_dataset()}}, \code{\link[rmorie:morie_arsau_analyze_individual_records]{morie_arsau_analyze_individual_records()}}, \code{\link[rmorie:morie_arsau_analyze_main_records]{morie_arsau_analyze_main_records()}}, \code{\link[rmorie:morie_arsau_analyze_probe_cycle_records]{morie_arsau_analyze_probe_cycle_records()}}, \code{\link[rmorie:morie_arsau_analyze_weapon_records]{morie_arsau_analyze_weapon_records()}}, \code{\link[rmorie:morie_arsau_available_datasets]{morie_arsau_available_datasets()}}, \code{\link[rmorie:morie_arsau_available_years]{morie_arsau_available_years()}}, \code{\link[rmorie:morie_arsau_ckan_url]{morie_arsau_ckan_url()}}, \code{\link[rmorie:morie_arsau_describe]{morie_arsau_describe()}}, \code{\link[rmorie:morie_arsau_download]{morie_arsau_download()}}, \code{\link[rmorie:morie_arsau_fetch_sidecar]{morie_arsau_fetch_sidecar()}}, \code{\link[rmorie:morie_arsau_load_aggregate_summary]{morie_arsau_load_aggregate_summary()}}, \code{\link[rmorie:morie_arsau_load_detailed_dataset]{morie_arsau_load_detailed_dataset()}}, \code{\link[rmorie:morie_arsau_load_individual_records]{morie_arsau_load_individual_records()}}, \code{\link[rmorie:morie_arsau_load_main_records]{morie_arsau_load_main_records()}}, \code{\link[rmorie:morie_arsau_load_probe_cycle_records]{morie_arsau_load_probe_cycle_records()}}, \code{\link[rmorie:morie_arsau_load_weapon_records]{morie_arsau_load_weapon_records()}}, \code{\link[rmorie:morie_arsau_read_markdown_dictionary]{morie_arsau_read_markdown_dictionary()}}, \code{\link[rmorie:morie_arsau_read_sidecar]{morie_arsau_read_sidecar()}}, \code{\link[rmorie:morie_arsau_read_xlsx_dictionary]{morie_arsau_read_xlsx_dictionary()}}, \code{\link[rmorie:morie_arsau_registry_df]{morie_arsau_registry_df()}}, \code{\link[rmorie:morie_arsau_sidecar_schema]{morie_arsau_sidecar_schema()}}, \code{\link[rmorie:morie_arsau_sidecar_to_frame]{morie_arsau_sidecar_to_frame()}}, \code{\link[rmorie:morie_calculate_ebac]{morie_calculate_ebac()}}, \code{\link[rmorie:morie_ckan_search]{morie_ckan_search()}}, \code{\link[rmorie:morie_cpads_canonicalize_frame]{morie_cpads_canonicalize_frame()}}, \code{\link[rmorie:morie_cpads_contract]{morie_cpads_contract()}}, \code{\link[rmorie:morie_cpads_has_raw_columns]{morie_cpads_has_raw_columns()}}, \code{\link[rmorie:morie_cpads_infer_file_format]{morie_cpads_infer_file_format()}}, \code{\link[rmorie:morie_cpads_missing_variables]{morie_cpads_missing_variables()}}, \code{\link[rmorie:morie_cpads_validate_frame]{morie_cpads_validate_frame()}}, \code{\link[rmorie:morie_dataset_catalog]{morie_dataset_catalog()}}, \code{\link[rmorie:morie_dataset_column_profile]{morie_dataset_column_profile()}}, \code{\link[rmorie:morie_dataset_detect_role]{morie_dataset_detect_role()}}, \code{\link[rmorie:morie_dataset_infer_level]{morie_dataset_infer_level()}}, \code{\link[rmorie:morie_dataset_load]{morie_dataset_load()}}, \code{\link[rmorie:morie_dataset_portal_catalog]{morie_dataset_portal_catalog()}}, \code{\link[rmorie:morie_dataset_portal_catalog_clear_cache]{morie_dataset_portal_catalog_clear_cache()}}, \code{\link[rmorie:morie_dataset_profile]{morie_dataset_profile()}}, \code{\link[rmorie:morie_dataset_profile_summary_table]{morie_dataset_profile_summary_table()}}, \code{\link[rmorie:morie_dataset_profile_to_list]{morie_dataset_profile_to_list()}}, \code{\link[rmorie:morie_dataset_suggest_plan]{morie_dataset_suggest_plan()}}, \code{\link[rmorie:morie_dataset_summarize_column]{morie_dataset_summarize_column()}}, \code{\link[rmorie:morie_datasets_arcgis_item_by_id]{morie_datasets_arcgis_item_by_id()}}, \code{\link[rmorie:morie_datasets_arcgis_item_metadata]{morie_datasets_arcgis_item_metadata()}}, \code{\link[rmorie:morie_datasets_arsau_aggregate_summary]{morie_datasets_arsau_aggregate_summary()}}, \code{\link[rmorie:morie_datasets_arsau_detailed_dataset]{morie_datasets_arsau_detailed_dataset()}}, \code{\link[rmorie:morie_datasets_arsau_uof_individual_records]{morie_datasets_arsau_uof_individual_records()}}, \code{\link[rmorie:morie_datasets_arsau_uof_main_records]{morie_datasets_arsau_uof_main_records()}}, \code{\link[rmorie:morie_datasets_arsau_uof_probe_cycle_records]{morie_datasets_arsau_uof_probe_cycle_records()}}, \code{\link[rmorie:morie_datasets_arsau_uof_weapon_records]{morie_datasets_arsau_uof_weapon_records()}}, \code{\link[rmorie:morie_datasets_bigquery]{morie_datasets_bigquery()}}, \code{\link[rmorie:morie_datasets_browse]{morie_datasets_browse()}}, \code{\link[rmorie:morie_datasets_calgary_community_crime_stats]{morie_datasets_calgary_community_crime_stats()}}, \code{\link[rmorie:morie_datasets_calgary_fire_response_calls]{morie_datasets_calgary_fire_response_calls()}}, \code{\link[rmorie:morie_datasets_calgary_fire_stations]{morie_datasets_calgary_fire_stations()}}, \code{\link[rmorie:morie_datasets_calgary_open_crime_adjacent_layers]{morie_datasets_calgary_open_crime_adjacent_layers()}}, \code{\link[rmorie:morie_datasets_calgary_opendata_bulk_layers]{morie_datasets_calgary_opendata_bulk_layers()}}, \code{\link[rmorie:morie_datasets_calgary_socrata_by_id]{morie_datasets_calgary_socrata_by_id()}}, \code{\link[rmorie:morie_datasets_chicago_arrests]{morie_datasets_chicago_arrests()}}, \code{\link[rmorie:morie_datasets_chicago_community_areas]{morie_datasets_chicago_community_areas()}}, \code{\link[rmorie:morie_datasets_chicago_crime]{morie_datasets_chicago_crime()}}, \code{\link[rmorie:morie_datasets_chicago_crime_map]{morie_datasets_chicago_crime_map()}}, \code{\link[rmorie:morie_datasets_chicago_crime_odata]{morie_datasets_chicago_crime_odata()}}, \code{\link[rmorie:morie_datasets_chicago_crime_resolved]{morie_datasets_chicago_crime_resolved()}}, \code{\link[rmorie:morie_datasets_chicago_crime_soql]{morie_datasets_chicago_crime_soql()}}, \code{\link[rmorie:morie_datasets_chicago_iucr_codes]{morie_datasets_chicago_iucr_codes()}}, \code{\link[rmorie:morie_datasets_chicago_neighborhoods]{morie_datasets_chicago_neighborhoods()}}, \code{\link[rmorie:morie_datasets_chicago_opendata_bulk_layers]{morie_datasets_chicago_opendata_bulk_layers()}}, \code{\link[rmorie:morie_datasets_chicago_police_beats]{morie_datasets_chicago_police_beats()}}, \code{\link[rmorie:morie_datasets_chicago_police_districts]{morie_datasets_chicago_police_districts()}}, \code{\link[rmorie:morie_datasets_chicago_socrata_by_id]{morie_datasets_chicago_socrata_by_id()}}, \code{\link[rmorie:morie_datasets_chicago_wards]{morie_datasets_chicago_wards()}}, \code{\link[rmorie:morie_datasets_ckan_package]{morie_datasets_ckan_package()}}, \code{\link[rmorie:morie_datasets_ckan_search]{morie_datasets_ckan_search()}}, \code{\link[rmorie:morie_datasets_cpads]{morie_datasets_cpads()}}, \code{\link[rmorie:morie_datasets_cpd_public_arrests]{morie_datasets_cpd_public_arrests()}}, \code{\link[rmorie:morie_datasets_edmonton_fire_stations]{morie_datasets_edmonton_fire_stations()}}, \code{\link[rmorie:morie_datasets_edmonton_open_crime_adjacent_layers]{morie_datasets_edmonton_open_crime_adjacent_layers()}}, \code{\link[rmorie:morie_datasets_edmonton_opendata_bulk_layers]{morie_datasets_edmonton_opendata_bulk_layers()}}, \code{\link[rmorie:morie_datasets_edmonton_police_stations]{morie_datasets_edmonton_police_stations()}}, \code{\link[rmorie:morie_datasets_edmonton_socrata_by_id]{morie_datasets_edmonton_socrata_by_id()}}, \code{\link[rmorie:morie_datasets_external_socrata_layers]{morie_datasets_external_socrata_layers()}}, \code{\link[rmorie:morie_datasets_load_by_key]{morie_datasets_load_by_key()}}, \code{\link[rmorie:morie_datasets_montreal_ckan_resource]{morie_datasets_montreal_ckan_resource()}}, \code{\link[rmorie:morie_datasets_montreal_justice_safety_layers]{morie_datasets_montreal_justice_safety_layers()}}, \code{\link[rmorie:morie_datasets_montreal_opendata_bulk_layers]{morie_datasets_montreal_opendata_bulk_layers()}}, \code{\link[rmorie:morie_datasets_montreal_sim_intervention_types]{morie_datasets_montreal_sim_intervention_types()}}, \code{\link[rmorie:morie_datasets_montreal_sim_interventions]{morie_datasets_montreal_sim_interventions()}}, \code{\link[rmorie:morie_datasets_namus_missing_persons]{morie_datasets_namus_missing_persons()}}, \code{\link[rmorie:morie_datasets_nibrs]{morie_datasets_nibrs()}}, \code{\link[rmorie:morie_datasets_nist_rds]{morie_datasets_nist_rds()}}, \code{\link[rmorie:morie_datasets_nyc_boroughs]{morie_datasets_nyc_boroughs()}}, \code{\link[rmorie:morie_datasets_nyc_boundaries_catalog]{morie_datasets_nyc_boundaries_catalog()}}, \code{\link[rmorie:morie_datasets_nyc_community_districts]{morie_datasets_nyc_community_districts()}}, \code{\link[rmorie:morie_datasets_nyc_council_districts]{morie_datasets_nyc_council_districts()}}, \code{\link[rmorie:morie_datasets_nyc_ntas_2020]{morie_datasets_nyc_ntas_2020()}}, \code{\link[rmorie:morie_datasets_nyc_nypd_arrests_historic]{morie_datasets_nyc_nypd_arrests_historic()}}, \code{\link[rmorie:morie_datasets_nyc_nypd_arrests_ytd]{morie_datasets_nyc_nypd_arrests_ytd()}}, \code{\link[rmorie:morie_datasets_nyc_nypd_boro_crosswalk]{morie_datasets_nyc_nypd_boro_crosswalk()}}, \code{\link[rmorie:morie_datasets_nyc_nypd_by_key]{morie_datasets_nyc_nypd_by_key()}}, \code{\link[rmorie:morie_datasets_nyc_nypd_complaint_historic]{morie_datasets_nyc_nypd_complaint_historic()}}, \code{\link[rmorie:morie_datasets_nyc_nypd_complaint_ytd]{morie_datasets_nyc_nypd_complaint_ytd()}}, \code{\link[rmorie:morie_datasets_nyc_nypd_hate_crimes]{morie_datasets_nyc_nypd_hate_crimes()}}, \code{\link[rmorie:morie_datasets_nyc_nypd_law_books]{morie_datasets_nyc_nypd_law_books()}}, \code{\link[rmorie:morie_datasets_nyc_nypd_layers]{morie_datasets_nyc_nypd_layers()}}, \code{\link[rmorie:morie_datasets_nyc_nypd_offense_codes]{morie_datasets_nyc_nypd_offense_codes()}}, \code{\link[rmorie:morie_datasets_nyc_nypd_resolved]{morie_datasets_nyc_nypd_resolved()}}, \code{\link[rmorie:morie_datasets_nyc_nypd_uof_incidents]{morie_datasets_nyc_nypd_uof_incidents()}}, \code{\link[rmorie:morie_datasets_nyc_nypd_uof_subjects]{morie_datasets_nyc_nypd_uof_subjects()}}, \code{\link[rmorie:morie_datasets_nyc_nypd_vehicle_stops]{morie_datasets_nyc_nypd_vehicle_stops()}}, \code{\link[rmorie:morie_datasets_nyc_opendata_bulk_layers]{morie_datasets_nyc_opendata_bulk_layers()}}, \code{\link[rmorie:morie_datasets_nyc_police_precincts]{morie_datasets_nyc_police_precincts()}}, \code{\link[rmorie:morie_datasets_nyc_school_districts]{morie_datasets_nyc_school_districts()}}, \code{\link[rmorie:morie_datasets_nyc_socrata_by_id]{morie_datasets_nyc_socrata_by_id()}}, \code{\link[rmorie:morie_datasets_nyc_stop_and_frisk]{morie_datasets_nyc_stop_and_frisk()}}, \code{\link[rmorie:morie_datasets_nyc_zctas]{morie_datasets_nyc_zctas()}}, \code{\link[rmorie:morie_datasets_ontario_ckan_by_key]{morie_datasets_ontario_ckan_by_key()}}, \code{\link[rmorie:morie_datasets_ontario_ckan_layers]{morie_datasets_ontario_ckan_layers()}}, \code{\link[rmorie:morie_datasets_otis_a01]{morie_datasets_otis_a01()}}, \code{\link[rmorie:morie_datasets_otis_a01_restrictive_confinement]{morie_datasets_otis_a01_restrictive_confinement()}}, \code{\link[rmorie:morie_datasets_otis_b01_segregation_detailed]{morie_datasets_otis_b01_segregation_detailed()}}, \code{\link[rmorie:morie_datasets_otis_b02_segregation_total_days]{morie_datasets_otis_b02_segregation_total_days()}}, \code{\link[rmorie:morie_datasets_otis_b03_seg_alerts_by_institution]{morie_datasets_otis_b03_seg_alerts_by_institution()}}, \code{\link[rmorie:morie_datasets_otis_b04_seg_consecutive_by_region]{morie_datasets_otis_b04_seg_consecutive_by_region()}}, \code{\link[rmorie:morie_datasets_otis_b05_seg_consecutive_lengths]{morie_datasets_otis_b05_seg_consecutive_lengths()}}, \code{\link[rmorie:morie_datasets_otis_b06_seg_reason_by_institution]{morie_datasets_otis_b06_seg_reason_by_institution()}}, \code{\link[rmorie:morie_datasets_otis_b07_seg_alerts_by_gender]{morie_datasets_otis_b07_seg_alerts_by_gender()}}, \code{\link[rmorie:morie_datasets_otis_b08_seg_consecutive_by_institution]{morie_datasets_otis_b08_seg_consecutive_by_institution()}}, \code{\link[rmorie:morie_datasets_otis_b09_seg_n_times]{morie_datasets_otis_b09_seg_n_times()}}, \code{\link[rmorie:morie_datasets_otis_c01_individuals_total]{morie_datasets_otis_c01_individuals_total()}}, \code{\link[rmorie:morie_datasets_otis_c02_individuals_by_institution]{morie_datasets_otis_c02_individuals_by_institution()}}, \code{\link[rmorie:morie_datasets_otis_c03_individuals_race_by_gender]{morie_datasets_otis_c03_individuals_race_by_gender()}}, \code{\link[rmorie:morie_datasets_otis_c04_individuals_race_by_region]{morie_datasets_otis_c04_individuals_race_by_region()}}, \code{\link[rmorie:morie_datasets_otis_c05_individuals_religion_by_region]{morie_datasets_otis_c05_individuals_religion_by_region()}}, \code{\link[rmorie:morie_datasets_otis_c06_individuals_age_by_region]{morie_datasets_otis_c06_individuals_age_by_region()}}, \code{\link[rmorie:morie_datasets_otis_c07_individuals_alerts]{morie_datasets_otis_c07_individuals_alerts()}}, \code{\link[rmorie:morie_datasets_otis_c08_individuals_religion_by_gender]{morie_datasets_otis_c08_individuals_religion_by_gender()}}, \code{\link[rmorie:morie_datasets_otis_c09_individuals_age_by_gender]{morie_datasets_otis_c09_individuals_age_by_gender()}}, \code{\link[rmorie:morie_datasets_otis_c10_aggregate_durations_by_institution]{morie_datasets_otis_c10_aggregate_durations_by_institution()}}, \code{\link[rmorie:morie_datasets_otis_c11_aggregate_lengths]{morie_datasets_otis_c11_aggregate_lengths()}}, \code{\link[rmorie:morie_datasets_otis_c12_aggregate_durations_by_region]{morie_datasets_otis_c12_aggregate_durations_by_region()}}, \code{\link[rmorie:morie_datasets_otis_d01_deaths_in_custody]{morie_datasets_otis_d01_deaths_in_custody()}}, \code{\link[rmorie:morie_datasets_otis_d02_deaths_by_gender]{morie_datasets_otis_d02_deaths_by_gender()}}, \code{\link[rmorie:morie_datasets_otis_d03_deaths_by_race]{morie_datasets_otis_d03_deaths_by_race()}}, \code{\link[rmorie:morie_datasets_otis_d04_deaths_by_religion]{morie_datasets_otis_d04_deaths_by_religion()}}, \code{\link[rmorie:morie_datasets_otis_d05_deaths_by_age_category]{morie_datasets_otis_d05_deaths_by_age_category()}}, \code{\link[rmorie:morie_datasets_otis_d06_cause_by_alert]{morie_datasets_otis_d06_cause_by_alert()}}, \code{\link[rmorie:morie_datasets_otis_d07_alerts_by_housing_unit]{morie_datasets_otis_d07_alerts_by_housing_unit()}}, \code{\link[rmorie:morie_datasets_ottawa_open_crime_adjacent_layers]{morie_datasets_ottawa_open_crime_adjacent_layers()}}, \code{\link[rmorie:morie_datasets_ottawa_opendata_bulk_layers]{morie_datasets_ottawa_opendata_bulk_layers()}}, \code{\link[rmorie:morie_datasets_siu_director_reports]{morie_datasets_siu_director_reports()}}, \code{\link[rmorie:morie_datasets_siu_report_fields]{morie_datasets_siu_report_fields()}}, \code{\link[rmorie:morie_datasets_siu_report_text]{morie_datasets_siu_report_text()}}, \code{\link[rmorie:morie_datasets_statcan_ccjs_cubes]{morie_datasets_statcan_ccjs_cubes()}}, \code{\link[rmorie:morie_datasets_statcan_cube_metadata]{morie_datasets_statcan_cube_metadata()}}, \code{\link[rmorie:morie_datasets_statcan_full_csv_url]{morie_datasets_statcan_full_csv_url()}}, \code{\link[rmorie:morie_datasets_statcan_vectors]{morie_datasets_statcan_vectors()}}, \code{\link[rmorie:morie_datasets_summary]{morie_datasets_summary()}}, \code{\link[rmorie:morie_datasets_toronto_ambulance_stations]{morie_datasets_toronto_ambulance_stations()}}, \code{\link[rmorie:morie_datasets_toronto_asr_miscellaneous]{morie_datasets_toronto_asr_miscellaneous()}}, \code{\link[rmorie:morie_datasets_toronto_open_ckan_resource]{morie_datasets_toronto_open_ckan_resource()}}, \code{\link[rmorie:morie_datasets_toronto_open_crime_adjacent_layers]{morie_datasets_toronto_open_crime_adjacent_layers()}}, \code{\link[rmorie:morie_datasets_toronto_opendata_bulk_layers]{morie_datasets_toronto_opendata_bulk_layers()}}, \code{\link[rmorie:morie_datasets_toronto_zoning_per_neighbourhood]{morie_datasets_toronto_zoning_per_neighbourhood()}}, \code{\link[rmorie:morie_datasets_tps_2008_firs]{morie_datasets_tps_2008_firs()}}, \code{\link[rmorie:morie_datasets_tps_2009_firs]{morie_datasets_tps_2009_firs()}}, \code{\link[rmorie:morie_datasets_tps_2010_firs]{morie_datasets_tps_2010_firs()}}, \code{\link[rmorie:morie_datasets_tps_2011_firs]{morie_datasets_tps_2011_firs()}}, \code{\link[rmorie:morie_datasets_tps_2012_firs]{morie_datasets_tps_2012_firs()}}, \code{\link[rmorie:morie_datasets_tps_2013_firs]{morie_datasets_tps_2013_firs()}}, \code{\link[rmorie:morie_datasets_tps_administrative]{morie_datasets_tps_administrative()}}, \code{\link[rmorie:morie_datasets_tps_arcgis_hub_by_id]{morie_datasets_tps_arcgis_hub_by_id()}}, \code{\link[rmorie:morie_datasets_tps_arcgis_hub_download]{morie_datasets_tps_arcgis_hub_download()}}, \code{\link[rmorie:morie_datasets_tps_arcgis_hub_layers]{morie_datasets_tps_arcgis_hub_layers()}}, \code{\link[rmorie:morie_datasets_tps_arrested_and_charged_persons]{morie_datasets_tps_arrested_and_charged_persons()}}, \code{\link[rmorie:morie_datasets_tps_arrests_and_strip_searches]{morie_datasets_tps_arrests_and_strip_searches()}}, \code{\link[rmorie:morie_datasets_tps_assault]{morie_datasets_tps_assault()}}, \code{\link[rmorie:morie_datasets_tps_automobile_ksi]{morie_datasets_tps_automobile_ksi()}}, \code{\link[rmorie:morie_datasets_tps_autotheft]{morie_datasets_tps_autotheft()}}, \code{\link[rmorie:morie_datasets_tps_bicycle_thefts]{morie_datasets_tps_bicycle_thefts()}}, \code{\link[rmorie:morie_datasets_tps_bicycletheft]{morie_datasets_tps_bicycletheft()}}, \code{\link[rmorie:morie_datasets_tps_breakandenter]{morie_datasets_tps_breakandenter()}}, \code{\link[rmorie:morie_datasets_tps_budget_2020]{morie_datasets_tps_budget_2020()}}, \code{\link[rmorie:morie_datasets_tps_budget_2021]{morie_datasets_tps_budget_2021()}}, \code{\link[rmorie:morie_datasets_tps_budget_2022]{morie_datasets_tps_budget_2022()}}, \code{\link[rmorie:morie_datasets_tps_budget_2023]{morie_datasets_tps_budget_2023()}}, \code{\link[rmorie:morie_datasets_tps_budget_2024]{morie_datasets_tps_budget_2024()}}, \code{\link[rmorie:morie_datasets_tps_budget_2025]{morie_datasets_tps_budget_2025()}}, \code{\link[rmorie:morie_datasets_tps_budget_2026]{morie_datasets_tps_budget_2026()}}, \code{\link[rmorie:morie_datasets_tps_budget_by_command]{morie_datasets_tps_budget_by_command()}}, \code{\link[rmorie:morie_datasets_tps_calls_for_service_attended]{morie_datasets_tps_calls_for_service_attended()}}, \code{\link[rmorie:morie_datasets_tps_community_safety_indicators]{morie_datasets_tps_community_safety_indicators()}}, \code{\link[rmorie:morie_datasets_tps_complaint_dispositions]{morie_datasets_tps_complaint_dispositions()}}, \code{\link[rmorie:morie_datasets_tps_cyclist_ksi]{morie_datasets_tps_cyclist_ksi()}}, \code{\link[rmorie:morie_datasets_tps_dispatched_calls_by_division]{morie_datasets_tps_dispatched_calls_by_division()}}, \code{\link[rmorie:morie_datasets_tps_facilities]{morie_datasets_tps_facilities()}}, \code{\link[rmorie:morie_datasets_tps_fatals_ksi]{morie_datasets_tps_fatals_ksi()}}, \code{\link[rmorie:morie_datasets_tps_firearms_top_calibres]{morie_datasets_tps_firearms_top_calibres()}}, \code{\link[rmorie:morie_datasets_tps_gross_expenditures_by_division]{morie_datasets_tps_gross_expenditures_by_division()}}, \code{\link[rmorie:morie_datasets_tps_gross_operating_budget]{morie_datasets_tps_gross_operating_budget()}}, \code{\link[rmorie:morie_datasets_tps_hatecrimes]{morie_datasets_tps_hatecrimes()}}, \code{\link[rmorie:morie_datasets_tps_homicide]{morie_datasets_tps_homicide()}}, \code{\link[rmorie:morie_datasets_tps_homicides]{morie_datasets_tps_homicides()}}, \code{\link[rmorie:morie_datasets_tps_intimate_partner_family_violence]{morie_datasets_tps_intimate_partner_family_violence()}}, \code{\link[rmorie:morie_datasets_tps_investigated_alleged_misconduct]{morie_datasets_tps_investigated_alleged_misconduct()}}, \code{\link[rmorie:morie_datasets_tps_killed_and_seriously_injured]{morie_datasets_tps_killed_and_seriously_injured()}}, \code{\link[rmorie:morie_datasets_tps_layers]{morie_datasets_tps_layers()}}, \code{\link[rmorie:morie_datasets_tps_major_crime]{morie_datasets_tps_major_crime()}}, \code{\link[rmorie:morie_datasets_tps_mha_apprehensions]{morie_datasets_tps_mha_apprehensions()}}, \code{\link[rmorie:morie_datasets_tps_miscellaneous_calls_for_service]{morie_datasets_tps_miscellaneous_calls_for_service()}}, \code{\link[rmorie:morie_datasets_tps_miscellaneous_data]{morie_datasets_tps_miscellaneous_data()}}, \code{\link[rmorie:morie_datasets_tps_miscellaneous_firearms]{morie_datasets_tps_miscellaneous_firearms()}}, \code{\link[rmorie:morie_datasets_tps_motorcylist_ksi]{morie_datasets_tps_motorcylist_ksi()}}, \code{\link[rmorie:morie_datasets_tps_neighbourhood_crime_rates]{morie_datasets_tps_neighbourhood_crime_rates()}}, \code{\link[rmorie:morie_datasets_tps_passenger_ksi]{morie_datasets_tps_passenger_ksi()}}, \code{\link[rmorie:morie_datasets_tps_patrol_zone]{morie_datasets_tps_patrol_zone()}}, \code{\link[rmorie:morie_datasets_tps_pedestrian_ksi]{morie_datasets_tps_pedestrian_ksi()}}, \code{\link[rmorie:morie_datasets_tps_personnel_by_command]{morie_datasets_tps_personnel_by_command()}}, \code{\link[rmorie:morie_datasets_tps_personnel_by_rank]{morie_datasets_tps_personnel_by_rank()}}, \code{\link[rmorie:morie_datasets_tps_personnel_by_rank_by_division]{morie_datasets_tps_personnel_by_rank_by_division()}}, \code{\link[rmorie:morie_datasets_tps_persons_in_crisis_calls_for_service_attended]{morie_datasets_tps_persons_in_crisis_calls_for_service_attended()}}, \code{\link[rmorie:morie_datasets_tps_police_divisions]{morie_datasets_tps_police_divisions()}}, \code{\link[rmorie:morie_datasets_tps_psdp_resolved]{morie_datasets_tps_psdp_resolved()}}, \code{\link[rmorie:morie_datasets_tps_regulated_interactions]{morie_datasets_tps_regulated_interactions()}}, \code{\link[rmorie:morie_datasets_tps_reported_crimes]{morie_datasets_tps_reported_crimes()}}, \code{\link[rmorie:morie_datasets_tps_robbery]{morie_datasets_tps_robbery()}}, \code{\link[rmorie:morie_datasets_tps_search_of_persons]{morie_datasets_tps_search_of_persons()}}, \code{\link[rmorie:morie_datasets_tps_shooting_firearm_discharges]{morie_datasets_tps_shooting_firearm_discharges()}}, \code{\link[rmorie:morie_datasets_tps_shootings]{morie_datasets_tps_shootings()}}, \code{\link[rmorie:morie_datasets_tps_staffing_by_command]{morie_datasets_tps_staffing_by_command()}}, \code{\link[rmorie:morie_datasets_tps_theft_from_motor_vehicle]{morie_datasets_tps_theft_from_motor_vehicle()}}, \code{\link[rmorie:morie_datasets_tps_theft_over]{morie_datasets_tps_theft_over()}}, \code{\link[rmorie:morie_datasets_tps_tickets_issued]{morie_datasets_tps_tickets_issued()}}, \code{\link[rmorie:morie_datasets_tps_top_20_offences_of_firearm_seizures]{morie_datasets_tps_top_20_offences_of_firearm_seizures()}}, \code{\link[rmorie:morie_datasets_tps_total_public_complaints]{morie_datasets_tps_total_public_complaints()}}, \code{\link[rmorie:morie_datasets_tps_traffic_collisions]{morie_datasets_tps_traffic_collisions()}}, \code{\link[rmorie:morie_datasets_tps_use_of_force_call_for_service_types]{morie_datasets_tps_use_of_force_call_for_service_types()}}, \code{\link[rmorie:morie_datasets_tps_use_of_force_call_sources_by_month]{morie_datasets_tps_use_of_force_call_sources_by_month()}}, \code{\link[rmorie:morie_datasets_tps_use_of_force_gender_composition]{morie_datasets_tps_use_of_force_gender_composition()}}, \code{\link[rmorie:morie_datasets_tps_use_of_force_location_of_occurrences]{morie_datasets_tps_use_of_force_location_of_occurrences()}}, \code{\link[rmorie:morie_datasets_tps_use_of_force_occurrence_category]{morie_datasets_tps_use_of_force_occurrence_category()}}, \code{\link[rmorie:morie_datasets_tps_use_of_force_time_of_day_trends]{morie_datasets_tps_use_of_force_time_of_day_trends()}}, \code{\link[rmorie:morie_datasets_tps_use_of_force_use_of_force_types_and_perceived_weapons]{morie_datasets_tps_use_of_force_use_of_force_types_and_perceived_weapons()}}, \code{\link[rmorie:morie_datasets_tps_victims_of_crime]{morie_datasets_tps_victims_of_crime()}}, \code{\link[rmorie:morie_datasets_vancouver_community_centres]{morie_datasets_vancouver_community_centres()}}, \code{\link[rmorie:morie_datasets_vancouver_community_food_markets]{morie_datasets_vancouver_community_food_markets()}}, \code{\link[rmorie:morie_datasets_vancouver_disability_parking]{morie_datasets_vancouver_disability_parking()}}, \code{\link[rmorie:morie_datasets_vancouver_fire_halls]{morie_datasets_vancouver_fire_halls()}}, \code{\link[rmorie:morie_datasets_vancouver_graffiti]{morie_datasets_vancouver_graffiti()}}, \code{\link[rmorie:morie_datasets_vancouver_homeless_shelters]{morie_datasets_vancouver_homeless_shelters()}}, \code{\link[rmorie:morie_datasets_vancouver_noise_control_areas]{morie_datasets_vancouver_noise_control_areas()}}, \code{\link[rmorie:morie_datasets_vancouver_opendata_bulk_layers]{morie_datasets_vancouver_opendata_bulk_layers()}}, \code{\link[rmorie:morie_datasets_vancouver_opendata_by_id]{morie_datasets_vancouver_opendata_by_id()}}, \code{\link[rmorie:morie_datasets_vancouver_opendata_layers]{morie_datasets_vancouver_opendata_layers()}}, \code{\link[rmorie:morie_datasets_vancouver_property_use_inspection_districts]{morie_datasets_vancouver_property_use_inspection_districts()}}, \code{\link[rmorie:morie_datasets_vancouver_public_art]{morie_datasets_vancouver_public_art()}}, \code{\link[rmorie:morie_datasets_vpd_crime]{morie_datasets_vpd_crime()}}, \code{\link[rmorie:morie_datasets_vpd_legal_disclaimer]{morie_datasets_vpd_legal_disclaimer()}}, \code{\link[rmorie:morie_fairness_apply_profile]{morie_fairness_apply_profile()}}, \code{\link[rmorie:morie_fairness_average_odds_difference]{morie_fairness_average_odds_difference()}}, \code{\link[rmorie:morie_fairness_bias_amplification]{morie_fairness_bias_amplification()}}, \code{\link[rmorie:morie_fairness_city_profile]{morie_fairness_city_profile()}}, \code{\link[rmorie:morie_fairness_column_map]{morie_fairness_column_map()}}, \code{\link[rmorie:morie_fairness_demographic_parity]{morie_fairness_demographic_parity()}}, \code{\link[rmorie:morie_fairness_disparate_impact]{morie_fairness_disparate_impact()}}, \code{\link[rmorie:morie_fairness_equalized_odds]{morie_fairness_equalized_odds()}}, \code{\link[rmorie:morie_fairness_get_city]{morie_fairness_get_city()}}, \code{\link[rmorie:morie_fairness_gini]{morie_fairness_gini()}}, \code{\link[rmorie:morie_fairness_list_cities]{morie_fairness_list_cities()}}, \code{\link[rmorie:morie_fairness_noisy_or_detection]{morie_fairness_noisy_or_detection()}}, \code{\link[rmorie:morie_fairness_predpol_aggregate_areas]{morie_fairness_predpol_aggregate_areas()}}, \code{\link[rmorie:morie_fairness_predpol_calibration_audit]{morie_fairness_predpol_calibration_audit()}}, \code{\link[rmorie:morie_fairness_predpol_score_disparity]{morie_fairness_predpol_score_disparity()}}, \code{\link[rmorie:morie_fairness_predpol_temporal_audit]{morie_fairness_predpol_temporal_audit()}}, \code{\link[rmorie:morie_fairness_register_city]{morie_fairness_register_city()}}, \code{\link[rmorie:morie_fairness_simulate_biased_crime_data]{morie_fairness_simulate_biased_crime_data()}}, \code{\link[rmorie:morie_fairness_xai_ale]{morie_fairness_xai_ale()}}, \code{\link[rmorie:morie_fairness_xai_ceteris_paribus]{morie_fairness_xai_ceteris_paribus()}}, \code{\link[rmorie:morie_fairness_xai_partial_dependence]{morie_fairness_xai_partial_dependence()}}, \code{\link[rmorie:morie_fairness_xai_permutation_importance]{morie_fairness_xai_permutation_importance()}}, \code{\link[rmorie:morie_fairness_xai_shap_values]{morie_fairness_xai_shap_values()}}, \code{\link[rmorie:morie_fetch]{morie_fetch()}}, \code{\link[rmorie:morie_fetch_arcgis]{morie_fetch_arcgis()}}, \code{\link[rmorie:morie_fetch_siu]{morie_fetch_siu()}}, \code{\link[rmorie:morie_infer_measurement_level]{morie_infer_measurement_level()}}, \code{\link[rmorie:morie_ingest_bigquery_build_sql]{morie_ingest_bigquery_build_sql()}}, \code{\link[rmorie:morie_ingest_bigquery_query]{morie_ingest_bigquery_query()}}, \code{\link[rmorie:morie_ingest_bigquery_table]{morie_ingest_bigquery_table()}}, \code{\link[rmorie:morie_ingest_chicago_crime]{morie_ingest_chicago_crime()}}, \code{\link[rmorie:morie_ingest_chicago_crime_bigquery]{morie_ingest_chicago_crime_bigquery()}}, \code{\link[rmorie:morie_ingest_chicago_resources]{morie_ingest_chicago_resources()}}, \code{\link[rmorie:morie_ingest_chicago_socrata]{morie_ingest_chicago_socrata()}}, \code{\link[rmorie:morie_ingest_cihi_xlsx]{morie_ingest_cihi_xlsx()}}, \code{\link[rmorie:morie_ingest_ckan_fetch_package_csvs]{morie_ingest_ckan_fetch_package_csvs()}}, \code{\link[rmorie:morie_ingest_ckan_package_search]{morie_ingest_ckan_package_search()}}, \code{\link[rmorie:morie_ingest_ckan_package_show]{morie_ingest_ckan_package_show()}}, \code{\link[rmorie:morie_ingest_ckan_read_resource]{morie_ingest_ckan_read_resource()}}, \code{\link[rmorie:morie_ingest_ckan_resource_show]{morie_ingest_ckan_resource_show()}}, \code{\link[rmorie:morie_ingest_ckan_search_packages]{morie_ingest_ckan_search_packages()}}, \code{\link[rmorie:morie_ingest_forensics_namus_missing]{morie_ingest_forensics_namus_missing()}}, \code{\link[rmorie:morie_ingest_forensics_nibrs]{morie_ingest_forensics_nibrs()}}, \code{\link[rmorie:morie_ingest_forensics_nist_rds]{morie_ingest_forensics_nist_rds()}}, \code{\link[rmorie:morie_ingest_statcan_cansim]{morie_ingest_statcan_cansim()}}, \code{\link[rmorie:morie_ingest_statcan_csv]{morie_ingest_statcan_csv()}}, \code{\link[rmorie:morie_ingest_tps_feature_layer]{morie_ingest_tps_feature_layer()}}, \code{\link[rmorie:morie_ingest_tps_fetch]{morie_ingest_tps_fetch()}}, \code{\link[rmorie:morie_ingest_tps_layers]{morie_ingest_tps_layers()}}, \code{\link[rmorie:morie_is_over_legal_limit]{morie_is_over_legal_limit()}}, \code{\link[rmorie:morie_laniyonu_actuarial_risk_disparity]{morie_laniyonu_actuarial_risk_disparity()}}, \code{\link[rmorie:morie_laniyonu_gentrification_policing]{morie_laniyonu_gentrification_policing()}}, \code{\link[rmorie:morie_laniyonu_smi_force_disparity]{morie_laniyonu_smi_force_disparity()}}, \code{\link[rmorie:morie_otis_aipw_ate]{morie_otis_aipw_ate()}}, \code{\link[rmorie:morie_otis_aipw_superlearner]{morie_otis_aipw_superlearner()}}, \code{\link[rmorie:morie_otis_alert_state_combo]{morie_otis_alert_state_combo()}}, \code{\link[rmorie:morie_otis_all_analyses]{morie_otis_all_analyses()}}, \code{\link[rmorie:morie_otis_analyze_a01]{morie_otis_analyze_a01()}}, \code{\link[rmorie:morie_otis_analyze_a01_ruhela_alt_age]{morie_otis_analyze_a01_ruhela_alt_age()}}, \code{\link[rmorie:morie_otis_analyze_a01_ruhela_alt_gender]{morie_otis_analyze_a01_ruhela_alt_gender()}}, \code{\link[rmorie:morie_otis_analyze_a01_ruhela_alt_toronto]{morie_otis_analyze_a01_ruhela_alt_toronto()}}, \code{\link[rmorie:morie_otis_analyze_a01_ruhela_formulations]{morie_otis_analyze_a01_ruhela_formulations()}}, \code{\link[rmorie:morie_otis_analyze_a01_ruhela_per_year]{morie_otis_analyze_a01_ruhela_per_year()}}, \code{\link[rmorie:morie_otis_analyze_a01_ruhela_subgroup_female]{morie_otis_analyze_a01_ruhela_subgroup_female()}}, \code{\link[rmorie:morie_otis_analyze_a01_ruhela_subgroup_male]{morie_otis_analyze_a01_ruhela_subgroup_male()}}, \code{\link[rmorie:morie_otis_analyze_a01_with_csi_context]{morie_otis_analyze_a01_with_csi_context()}}, \code{\link[rmorie:morie_otis_analyze_all]{morie_otis_analyze_all()}}, \code{\link[rmorie:morie_otis_analyze_b01]{morie_otis_analyze_b01()}}, \code{\link[rmorie:morie_otis_analyze_b01_ruhela_alt_age]{morie_otis_analyze_b01_ruhela_alt_age()}}, \code{\link[rmorie:morie_otis_analyze_b01_ruhela_alt_gender]{morie_otis_analyze_b01_ruhela_alt_gender()}}, \code{\link[rmorie:morie_otis_analyze_b01_ruhela_alt_toronto]{morie_otis_analyze_b01_ruhela_alt_toronto()}}, \code{\link[rmorie:morie_otis_analyze_b01_ruhela_formulations]{morie_otis_analyze_b01_ruhela_formulations()}}, \code{\link[rmorie:morie_otis_analyze_b01_ruhela_per_year]{morie_otis_analyze_b01_ruhela_per_year()}}, \code{\link[rmorie:morie_otis_analyze_b01_ruhela_subgroup_female]{morie_otis_analyze_b01_ruhela_subgroup_female()}}, \code{\link[rmorie:morie_otis_analyze_b01_ruhela_subgroup_male]{morie_otis_analyze_b01_ruhela_subgroup_male()}}, \code{\link[rmorie:morie_otis_analyze_b02]{morie_otis_analyze_b02()}}, \code{\link[rmorie:morie_otis_analyze_b02_ruhela_alt_age]{morie_otis_analyze_b02_ruhela_alt_age()}}, \code{\link[rmorie:morie_otis_analyze_b02_ruhela_alt_region]{morie_otis_analyze_b02_ruhela_alt_region()}}, \code{\link[rmorie:morie_otis_analyze_b02_ruhela_formulations]{morie_otis_analyze_b02_ruhela_formulations()}}, \code{\link[rmorie:morie_otis_analyze_b03]{morie_otis_analyze_b03()}}, \code{\link[rmorie:morie_otis_analyze_b03_ruhela_aggregate]{morie_otis_analyze_b03_ruhela_aggregate()}}, \code{\link[rmorie:morie_otis_analyze_b04]{morie_otis_analyze_b04()}}, \code{\link[rmorie:morie_otis_analyze_b04_ruhela_aggregate]{morie_otis_analyze_b04_ruhela_aggregate()}}, \code{\link[rmorie:morie_otis_analyze_b05]{morie_otis_analyze_b05()}}, \code{\link[rmorie:morie_otis_analyze_b05_mandela_classification]{morie_otis_analyze_b05_mandela_classification()}}, \code{\link[rmorie:morie_otis_analyze_b05_ruhela_aggregate]{morie_otis_analyze_b05_ruhela_aggregate()}}, \code{\link[rmorie:morie_otis_analyze_b06]{morie_otis_analyze_b06()}}, \code{\link[rmorie:morie_otis_analyze_b06_ruhela_aggregate]{morie_otis_analyze_b06_ruhela_aggregate()}}, \code{\link[rmorie:morie_otis_analyze_b07]{morie_otis_analyze_b07()}}, \code{\link[rmorie:morie_otis_analyze_b07_ruhela_aggregate]{morie_otis_analyze_b07_ruhela_aggregate()}}, \code{\link[rmorie:morie_otis_analyze_b08]{morie_otis_analyze_b08()}}, \code{\link[rmorie:morie_otis_analyze_b08_ruhela_aggregate]{morie_otis_analyze_b08_ruhela_aggregate()}}, \code{\link[rmorie:morie_otis_analyze_b09]{morie_otis_analyze_b09()}}, \code{\link[rmorie:morie_otis_analyze_b09_ruhela_aggregate]{morie_otis_analyze_b09_ruhela_aggregate()}}, \code{\link[rmorie:morie_otis_analyze_c_chi2]{morie_otis_analyze_c_chi2()}}, \code{\link[rmorie:morie_otis_analyze_c01]{morie_otis_analyze_c01()}}, \code{\link[rmorie:morie_otis_analyze_c01_ruhela_aggregate]{morie_otis_analyze_c01_ruhela_aggregate()}}, \code{\link[rmorie:morie_otis_analyze_c01_ruhela_aggregate_region_cluster]{morie_otis_analyze_c01_ruhela_aggregate_region_cluster()}}, \code{\link[rmorie:morie_otis_analyze_c02]{morie_otis_analyze_c02()}}, \code{\link[rmorie:morie_otis_analyze_c02_ruhela_aggregate]{morie_otis_analyze_c02_ruhela_aggregate()}}, \code{\link[rmorie:morie_otis_analyze_c03]{morie_otis_analyze_c03()}}, \code{\link[rmorie:morie_otis_analyze_c03_ruhela_aggregate]{morie_otis_analyze_c03_ruhela_aggregate()}}, \code{\link[rmorie:morie_otis_analyze_c04_ruhela_aggregate]{morie_otis_analyze_c04_ruhela_aggregate()}}, \code{\link[rmorie:morie_otis_analyze_c04_ruhela_aggregate_region_cluster]{morie_otis_analyze_c04_ruhela_aggregate_region_cluster()}}, \code{\link[rmorie:morie_otis_analyze_c05_ruhela_aggregate]{morie_otis_analyze_c05_ruhela_aggregate()}}, \code{\link[rmorie:morie_otis_analyze_c06_ruhela_aggregate]{morie_otis_analyze_c06_ruhela_aggregate()}}, \code{\link[rmorie:morie_otis_analyze_c07]{morie_otis_analyze_c07()}}, \code{\link[rmorie:morie_otis_analyze_c07_ruhela_aggregate]{morie_otis_analyze_c07_ruhela_aggregate()}}, \code{\link[rmorie:morie_otis_analyze_c08_ruhela_aggregate]{morie_otis_analyze_c08_ruhela_aggregate()}}, \code{\link[rmorie:morie_otis_analyze_c09_ruhela_aggregate]{morie_otis_analyze_c09_ruhela_aggregate()}}, \code{\link[rmorie:morie_otis_analyze_c10]{morie_otis_analyze_c10()}}, \code{\link[rmorie:morie_otis_analyze_c10_ruhela_aggregate]{morie_otis_analyze_c10_ruhela_aggregate()}}, \code{\link[rmorie:morie_otis_analyze_c11]{morie_otis_analyze_c11()}}, \code{\link[rmorie:morie_otis_analyze_c11_mandela_classification]{morie_otis_analyze_c11_mandela_classification()}}, \code{\link[rmorie:morie_otis_analyze_c11_ruhela_aggregate]{morie_otis_analyze_c11_ruhela_aggregate()}}, \code{\link[rmorie:morie_otis_analyze_c12]{morie_otis_analyze_c12()}}, \code{\link[rmorie:morie_otis_analyze_c12_ruhela_aggregate]{morie_otis_analyze_c12_ruhela_aggregate()}}, \code{\link[rmorie:morie_otis_analyze_d_chi2]{morie_otis_analyze_d_chi2()}}, \code{\link[rmorie:morie_otis_analyze_d01]{morie_otis_analyze_d01()}}, \code{\link[rmorie:morie_otis_analyze_d02_ruhela_aggregate]{morie_otis_analyze_d02_ruhela_aggregate()}}, \code{\link[rmorie:morie_otis_analyze_d03_ruhela_aggregate]{morie_otis_analyze_d03_ruhela_aggregate()}}, \code{\link[rmorie:morie_otis_analyze_d04_ruhela_aggregate]{morie_otis_analyze_d04_ruhela_aggregate()}}, \code{\link[rmorie:morie_otis_analyze_d05_ruhela_aggregate]{morie_otis_analyze_d05_ruhela_aggregate()}}, \code{\link[rmorie:morie_otis_analyze_d06]{morie_otis_analyze_d06()}}, \code{\link[rmorie:morie_otis_analyze_d07]{morie_otis_analyze_d07()}}, \code{\link[rmorie:morie_otis_analyze_otis_mandela_provincial_vs_federal]{morie_otis_analyze_otis_mandela_provincial_vs_federal()}}, \code{\link[rmorie:morie_otis_analyze_ruhela_grid]{morie_otis_analyze_ruhela_grid()}}, \code{\link[rmorie:morie_otis_analyze_ruhela_master]{morie_otis_analyze_ruhela_master()}}, \code{\link[rmorie:morie_otis_analyze_ruhela_per_year]{morie_otis_analyze_ruhela_per_year()}}, \code{\link[rmorie:morie_otis_analyzers]{morie_otis_analyzers()}}, \code{\link[rmorie:morie_otis_astcmb]{morie_otis_astcmb()}}, \code{\link[rmorie:morie_otis_causal_grid]{morie_otis_causal_grid()}}, \code{\link[rmorie:morie_otis_churn_analyze_all]{morie_otis_churn_analyze_all()}}, \code{\link[rmorie:morie_otis_classify_mandela_combo]{morie_otis_classify_mandela_combo()}}, \code{\link[rmorie:morie_otis_descriptives]{morie_otis_descriptives()}}, \code{\link[rmorie:morie_otis_disciplinary_medical_overlap]{morie_otis_disciplinary_medical_overlap()}}, \code{\link[rmorie:morie_otis_dml]{morie_otis_dml()}}, \code{\link[rmorie:morie_otis_embedding_distribution]{morie_otis_embedding_distribution()}}, \code{\link[rmorie:morie_otis_intra_year_transition_matrix]{morie_otis_intra_year_transition_matrix()}}, \code{\link[rmorie:morie_otis_ipw_ate]{morie_otis_ipw_ate()}}, \code{\link[rmorie:morie_otis_irm_dml]{morie_otis_irm_dml()}}, \code{\link[rmorie:morie_otis_irr_glmm_vm]{morie_otis_irr_glmm_vm()}}, \code{\link[rmorie:morie_otis_load]{morie_otis_load()}}, \code{\link[rmorie:morie_otis_make_pair_a]{morie_otis_make_pair_a()}}, \code{\link[rmorie:morie_otis_make_pair_alert_to_volatility_a01]{morie_otis_make_pair_alert_to_volatility_a01()}}, \code{\link[rmorie:morie_otis_make_pair_alert_to_volatility_all]{morie_otis_make_pair_alert_to_volatility_all()}}, \code{\link[rmorie:morie_otis_make_pair_alert_to_volatility_naive]{morie_otis_make_pair_alert_to_volatility_naive()}}, \code{\link[rmorie:morie_otis_make_pair_alert_to_volatility_ruhela]{morie_otis_make_pair_alert_to_volatility_ruhela()}}, \code{\link[rmorie:morie_otis_make_pair_b]{morie_otis_make_pair_b()}}, \code{\link[rmorie:morie_otis_make_pair_c]{morie_otis_make_pair_c()}}, \code{\link[rmorie:morie_otis_mortification_cooccurrence]{morie_otis_mortification_cooccurrence()}}, \code{\link[rmorie:morie_otis_otdesc]{morie_otis_otdesc()}}, \code{\link[rmorie:morie_otis_otdml]{morie_otis_otdml()}}, \code{\link[rmorie:morie_otis_path_complexity_gini]{morie_otis_path_complexity_gini()}}, \code{\link[rmorie:morie_otis_plr]{morie_otis_plr()}}, \code{\link[rmorie:morie_otis_psm]{morie_otis_psm()}}, \code{\link[rmorie:morie_otis_psm_subclass]{morie_otis_psm_subclass()}}, \code{\link[rmorie:morie_otis_rc_trends]{morie_otis_rc_trends()}}, \code{\link[rmorie:morie_otis_rctrnd]{morie_otis_rctrnd()}}, \code{\link[rmorie:morie_otis_regC_demog_contingency]{morie_otis_regC_demog_contingency()}}, \code{\link[rmorie:morie_otis_region_alert_state_richness]{morie_otis_region_alert_state_richness()}}, \code{\link[rmorie:morie_otis_regional_placement]{morie_otis_regional_placement()}}, \code{\link[rmorie:morie_otis_repeat_placement_concentration]{morie_otis_repeat_placement_concentration()}}, \code{\link[rmorie:morie_otis_rplace]{morie_otis_rplace()}}, \code{\link[rmorie:morie_otis_tps_analyze_all]{morie_otis_tps_analyze_all()}}, \code{\link[rmorie:morie_otis_tps_composite_overlay]{morie_otis_tps_composite_overlay()}}, \code{\link[rmorie:morie_otis_tps_per_region_rollup]{morie_otis_tps_per_region_rollup()}}, \code{\link[rmorie:morie_otis_tps_yoy_correlation]{morie_otis_tps_yoy_correlation()}}, \code{\link[rmorie:morie_otis_volat]{morie_otis_volat()}}, \code{\link[rmorie:morie_otis_volatility]{morie_otis_volatility()}}, \code{\link[rmorie:morie_otis_within_year_placement_count]{morie_otis_within_year_placement_count()}}, \code{\link[rmorie:morie_otis_within_year_region_diversity]{morie_otis_within_year_region_diversity()}}, \code{\link[rmorie:morie_parse_nypd_law_code]{morie_parse_nypd_law_code()}}, \code{\link[rmorie:morie_predpol_aggregate_areas]{morie_predpol_aggregate_areas()}}, \code{\link[rmorie:morie_predpol_calibration_audit]{morie_predpol_calibration_audit()}}, \code{\link[rmorie:morie_predpol_score_disparity]{morie_predpol_score_disparity()}}, \code{\link[rmorie:morie_profile_dataset]{morie_profile_dataset()}}, \code{\link[rmorie:morie_siu_all_analyses]{morie_siu_all_analyses()}}, \code{\link[rmorie:morie_siu_anomaly_check]{morie_siu_anomaly_check()}}, \code{\link[rmorie:morie_siu_audit_case]{morie_siu_audit_case()}}, \code{\link[rmorie:morie_siu_audit_columns]{morie_siu_audit_columns()}}, \code{\link[rmorie:morie_siu_by_police_service]{morie_siu_by_police_service()}}, \code{\link[rmorie:morie_siu_by_year]{morie_siu_by_year()}}, \code{\link[rmorie:morie_siu_cache_path]{morie_siu_cache_path()}}, \code{\link[rmorie:morie_siu_case_counts]{morie_siu_case_counts()}}, \code{\link[rmorie:morie_siu_charges_by_year_chi2]{morie_siu_charges_by_year_chi2()}}, \code{\link[rmorie:morie_siu_compare]{morie_siu_compare()}}, \code{\link[rmorie:morie_siu_decision_timing]{morie_siu_decision_timing()}}, \code{\link[rmorie:morie_siu_demographics]{morie_siu_demographics()}}, \code{\link[rmorie:morie_siu_fetch_cases]{morie_siu_fetch_cases()}}, \code{\link[rmorie:morie_siu_fetch_dataframe]{morie_siu_fetch_dataframe()}}, \code{\link[rmorie:morie_siu_index]{morie_siu_index()}}, \code{\link[rmorie:morie_siu_index_url]{morie_siu_index_url()}}, \code{\link[rmorie:morie_siu_llm_extract]{morie_siu_llm_extract()}}, \code{\link[rmorie:morie_siu_mental_health_race_indicators]{morie_siu_mental_health_race_indicators()}}, \code{\link[rmorie:morie_siu_parse_html]{morie_siu_parse_html()}}, \code{\link[rmorie:morie_siu_parse_news_html]{morie_siu_parse_news_html()}}, \code{\link[rmorie:morie_siu_record_correction]{morie_siu_record_correction()}}, \code{\link[rmorie:morie_siu_refresh_manifest]{morie_siu_refresh_manifest()}}, \code{\link[rmorie:morie_siu_sanity_check]{morie_siu_sanity_check()}}, \code{\link[rmorie:morie_siu_translate]{morie_siu_translate()}}, \code{\link[rmorie:morie_siu_translate_fr_to_en]{morie_siu_translate_fr_to_en()}}, \code{\link[rmorie:morie_siuiap_cite]{morie_siuiap_cite()}}, \code{\link[rmorie:morie_siuiap_panel_summary]{morie_siuiap_panel_summary()}}, \code{\link[rmorie:morie_spatial_voting_aldrich_mckelvey]{morie_spatial_voting_aldrich_mckelvey()}}, \code{\link[rmorie:morie_spatial_voting_alpha_nominate]{morie_spatial_voting_alpha_nominate()}}, \code{\link[rmorie:morie_spatial_voting_anchoring_vignettes]{morie_spatial_voting_anchoring_vignettes()}}, \code{\link[rmorie:morie_spatial_voting_bayesian_am]{morie_spatial_voting_bayesian_am()}}, \code{\link[rmorie:morie_spatial_voting_bayesian_irt_likelihood]{morie_spatial_voting_bayesian_irt_likelihood()}}, \code{\link[rmorie:morie_spatial_voting_bayesian_irt_posterior]{morie_spatial_voting_bayesian_irt_posterior()}}, \code{\link[rmorie:morie_spatial_voting_bayesian_mds]{morie_spatial_voting_bayesian_mds()}}, \code{\link[rmorie:morie_spatial_voting_bayesian_unfolding]{morie_spatial_voting_bayesian_unfolding()}}, \code{\link[rmorie:morie_spatial_voting_blackbox]{morie_spatial_voting_blackbox()}}, \code{\link[rmorie:morie_spatial_voting_cjr_irt]{morie_spatial_voting_cjr_irt()}}, \code{\link[rmorie:morie_spatial_voting_classical_mds]{morie_spatial_voting_classical_mds()}}, \code{\link[rmorie:morie_spatial_voting_cutting_lines]{morie_spatial_voting_cutting_lines()}}, \code{\link[rmorie:morie_spatial_voting_double_centering]{morie_spatial_voting_double_centering()}}, \code{\link[rmorie:morie_spatial_voting_dw_nominate]{morie_spatial_voting_dw_nominate()}}, \code{\link[rmorie:morie_spatial_voting_dynamic_irt]{morie_spatial_voting_dynamic_irt()}}, \code{\link[rmorie:morie_spatial_voting_em_irt]{morie_spatial_voting_em_irt()}}, \code{\link[rmorie:morie_spatial_voting_ideal_point_recovery]{morie_spatial_voting_ideal_point_recovery()}}, \code{\link[rmorie:morie_spatial_voting_indscal]{morie_spatial_voting_indscal()}}, \code{\link[rmorie:morie_spatial_voting_mds_fit_stats]{morie_spatial_voting_mds_fit_stats()}}, \code{\link[rmorie:morie_spatial_voting_mlsmu6]{morie_spatial_voting_mlsmu6()}}, \code{\link[rmorie:morie_spatial_voting_nominate_bootstrap]{morie_spatial_voting_nominate_bootstrap()}}, \code{\link[rmorie:morie_spatial_voting_nominate_loglik]{morie_spatial_voting_nominate_loglik()}}, \code{\link[rmorie:morie_spatial_voting_nominate_utility]{morie_spatial_voting_nominate_utility()}}, \code{\link[rmorie:morie_spatial_voting_nominate_vote_prob]{morie_spatial_voting_nominate_vote_prob()}}, \code{\link[rmorie:morie_spatial_voting_nonmetric_mds]{morie_spatial_voting_nonmetric_mds()}}, \code{\link[rmorie:morie_spatial_voting_nonparametric_bootstrap]{morie_spatial_voting_nonparametric_bootstrap()}}, \code{\link[rmorie:morie_spatial_voting_normal_vectors]{morie_spatial_voting_normal_vectors()}}, \code{\link[rmorie:morie_spatial_voting_optimal_classification]{morie_spatial_voting_optimal_classification()}}, \code{\link[rmorie:morie_spatial_voting_ordered_oc]{morie_spatial_voting_ordered_oc()}}, \code{\link[rmorie:morie_spatial_voting_ordinal_irt]{morie_spatial_voting_ordinal_irt()}}, \code{\link[rmorie:morie_spatial_voting_procrustes]{morie_spatial_voting_procrustes()}}, \code{\link[rmorie:morie_spatial_voting_smacof]{morie_spatial_voting_smacof()}}, \code{\link[rmorie:morie_spatial_voting_smacof_unfolding]{morie_spatial_voting_smacof_unfolding()}}, \code{\link[rmorie:morie_spatial_voting_unfolding_stress]{morie_spatial_voting_unfolding_stress()}}, \code{\link[rmorie:morie_spatial_voting_wordfish]{morie_spatial_voting_wordfish()}}, \code{\link[rmorie:morie_suggest_analysis_plan]{morie_suggest_analysis_plan()}}, \code{\link[rmorie:morie_synth_otis]{morie_synth_otis()}}, \code{\link[rmorie:morie_synth_otis_all]{morie_synth_otis_all()}}, \code{\link[rmorie:morie_to_hood_crosswalk]{morie_to_hood_crosswalk()}}, \code{\link[rmorie:morie_to_neighbourhoods]{morie_to_neighbourhoods()}}, \code{\link[rmorie:morie_tps_add_hood_140_from_158]{morie_tps_add_hood_140_from_158()}}, \code{\link[rmorie:morie_tps_add_hood_158_from_140]{morie_tps_add_hood_158_from_140()}}, \code{\link[rmorie:morie_tps_aggregate_158_to_140]{morie_tps_aggregate_158_to_140()}}, \code{\link[rmorie:morie_tps_analyze_all]{morie_tps_analyze_all()}}, \code{\link[rmorie:morie_tps_analyze_csi_from_dataframes]{morie_tps_analyze_csi_from_dataframes()}}, \code{\link[rmorie:morie_tps_analyze_one]{morie_tps_analyze_one()}}, \code{\link[rmorie:morie_tps_arima_forecast]{morie_tps_arima_forecast()}}, \code{\link[rmorie:morie_tps_assert_hood_version]{morie_tps_assert_hood_version()}}, \code{\link[rmorie:morie_tps_available_formats]{morie_tps_available_formats()}}, \code{\link[rmorie:morie_tps_bivariate_moran]{morie_tps_bivariate_moran()}}, \code{\link[rmorie:morie_tps_bivariate_morans_i]{morie_tps_bivariate_morans_i()}}, \code{\link[rmorie:morie_tps_category_correlation_matrix]{morie_tps_category_correlation_matrix()}}, \code{\link[rmorie:morie_tps_changepoint_detection]{morie_tps_changepoint_detection()}}, \code{\link[rmorie:morie_tps_compare_hawkes_kernels]{morie_tps_compare_hawkes_kernels()}}, \code{\link[rmorie:morie_tps_composite_index]{morie_tps_composite_index()}}, \code{\link[rmorie:morie_tps_crime_compare]{morie_tps_crime_compare()}}, \code{\link[rmorie:morie_tps_criminal_network_graph]{morie_tps_criminal_network_graph()}}, \code{\link[rmorie:morie_tps_csi_per_neighbourhood]{morie_tps_csi_per_neighbourhood()}}, \code{\link[rmorie:morie_tps_csi_per_year]{morie_tps_csi_per_year()}}, \code{\link[rmorie:morie_tps_csi_weight]{morie_tps_csi_weight()}}, \code{\link[rmorie:morie_tps_data_dir]{morie_tps_data_dir()}}, \code{\link[rmorie:morie_tps_dbscan_clusters]{morie_tps_dbscan_clusters()}}, \code{\link[rmorie:morie_tps_disaggregate_140_to_158]{morie_tps_disaggregate_140_to_158()}}, \code{\link[rmorie:morie_tps_district_for_centroid]{morie_tps_district_for_centroid()}}, \code{\link[rmorie:morie_tps_fetch_category]{morie_tps_fetch_category()}}, \code{\link[rmorie:morie_tps_fetch_dataframe]{morie_tps_fetch_dataframe()}}, \code{\link[rmorie:morie_tps_fokker_planck_grid]{morie_tps_fokker_planck_grid()}}, \code{\link[rmorie:morie_tps_getis_ord_g_star]{morie_tps_getis_ord_g_star()}}, \code{\link[rmorie:morie_tps_gini_concentration]{morie_tps_gini_concentration()}}, \code{\link[rmorie:morie_tps_hawkes_advanced_fit]{morie_tps_hawkes_advanced_fit()}}, \code{\link[rmorie:morie_tps_hawkes_markovian_vs_nonmarkovian]{morie_tps_hawkes_markovian_vs_nonmarkovian()}}, \code{\link[rmorie:morie_tps_hawkes_temporal_fit]{morie_tps_hawkes_temporal_fit()}}, \code{\link[rmorie:morie_tps_inspection_game_phase]{morie_tps_inspection_game_phase()}}, \code{\link[rmorie:morie_tps_kde_density]{morie_tps_kde_density()}}, \code{\link[rmorie:morie_tps_langevin_simulate]{morie_tps_langevin_simulate()}}, \code{\link[rmorie:morie_tps_levy_flight_alpha]{morie_tps_levy_flight_alpha()}}, \code{\link[rmorie:morie_tps_list_categories]{morie_tps_list_categories()}}, \code{\link[rmorie:morie_tps_list_datasets]{morie_tps_list_datasets()}}, \code{\link[rmorie:morie_tps_list_formats]{morie_tps_list_formats()}}, \code{\link[rmorie:morie_tps_load]{morie_tps_load()}}, \code{\link[rmorie:morie_tps_load_dataset]{morie_tps_load_dataset()}}, \code{\link[rmorie:morie_tps_local_morans_i]{morie_tps_local_morans_i()}}, \code{\link[rmorie:morie_tps_lotka_volterra_police_crime]{morie_tps_lotka_volterra_police_crime()}}, \code{\link[rmorie:morie_tps_moran_sweep_heatmap]{morie_tps_moran_sweep_heatmap()}}, \code{\link[rmorie:morie_tps_morans_i_neighbourhood]{morie_tps_morans_i_neighbourhood()}}, \code{\link[rmorie:morie_tps_neighbourhood_concentration]{morie_tps_neighbourhood_concentration()}}, \code{\link[rmorie:morie_tps_offence_summary]{morie_tps_offence_summary()}}, \code{\link[rmorie:morie_tps_polygon_morans_i]{morie_tps_polygon_morans_i()}}, \code{\link[rmorie:morie_tps_pretty_label]{morie_tps_pretty_label()}}, \code{\link[rmorie:morie_tps_project_xy]{morie_tps_project_xy()}}, \code{\link[rmorie:morie_tps_psdp_layers]{morie_tps_psdp_layers()}}, \code{\link[rmorie:morie_tps_render_choropleth]{morie_tps_render_choropleth()}}, \code{\link[rmorie:morie_tps_render_dbscan]{morie_tps_render_dbscan()}}, \code{\link[rmorie:morie_tps_render_district_proportional]{morie_tps_render_district_proportional()}}, \code{\link[rmorie:morie_tps_render_points]{morie_tps_render_points()}}, \code{\link[rmorie:morie_tps_render_quad]{morie_tps_render_quad()}}, \code{\link[rmorie:morie_tps_render_satscan_panel]{morie_tps_render_satscan_panel()}}, \code{\link[rmorie:morie_tps_render_yearly_grid]{morie_tps_render_yearly_grid()}}, \code{\link[rmorie:morie_tps_resolve_hood_col]{morie_tps_resolve_hood_col()}}, \code{\link[rmorie:morie_tps_ripley_k]{morie_tps_ripley_k()}}, \code{\link[rmorie:morie_tps_sarima_forecast]{morie_tps_sarima_forecast()}}, \code{\link[rmorie:morie_tps_sdb_reaction_diffusion]{morie_tps_sdb_reaction_diffusion()}}, \code{\link[rmorie:morie_tps_sdb_turing_demo]{morie_tps_sdb_turing_demo()}}, \code{\link[rmorie:morie_tps_seasonal_pattern]{morie_tps_seasonal_pattern()}}, \code{\link[rmorie:morie_tps_spatial_summary]{morie_tps_spatial_summary()}}, \code{\link[rmorie:morie_tps_statphysics_analyze_all]{morie_tps_statphysics_analyze_all()}}, \code{\link[rmorie:morie_tps_temporal_summary]{morie_tps_temporal_summary()}}, \code{\link[rmorie:morie_tps_urban_scaling_beta]{morie_tps_urban_scaling_beta()}}, \code{\link[rmorie:morie_tps_use_of_force]{morie_tps_use_of_force()}}, \code{\link[rmorie:morie_tps_year_over_year_trend]{morie_tps_year_over_year_trend()}}, \code{\link[rmorie:morie_tps_year_to_hood_version]{morie_tps_year_to_hood_version()}}, \code{\link[rmorie:morie_tps_yoy_panel]{morie_tps_yoy_panel()}}, \code{\link[rmorie:morie_vpd_download_instructions]{morie_vpd_download_instructions()}}, \code{\link[rmorie:mrm_classify_mandela]{mrm_classify_mandela()}}, \code{\link[rmorie:mrm_otis_mandela_spectrum]{mrm_otis_mandela_spectrum()}}, \code{\link[rmorie:mrm_otis_mortification_cooccurrence]{mrm_otis_mortification_cooccurrence()}}, \code{\link[rmorie:mrm_otis_placement_concentration]{mrm_otis_placement_concentration()}}, \code{\link[rmorie:mrm_otis_region_locality]{mrm_otis_region_locality()}}, \code{\link[rmorie:mrm_otis_seg_duration_km]{mrm_otis_seg_duration_km()}}, \code{\link[rmorie:mrm_siu_case_to_decision_km]{mrm_siu_case_to_decision_km()}}, \code{\link[rmorie:mrm_siu_outcome_classifier]{mrm_siu_outcome_classifier()}}, \code{\link[rmorie:mrm_siu_per_service_rate]{mrm_siu_per_service_rate()}}, \code{\link[rmorie:mrm_tps_levy_scaling]{mrm_tps_levy_scaling()}}, \code{\link[rmorie:mrm_tps_load_hawkes_refit]{mrm_tps_load_hawkes_refit()}}, \code{\link[rmorie:mrm_tps_moran_clustering]{mrm_tps_moran_clustering()}}, \code{\link[rmorie:mrm_tps_neighbourhood_recurrence_km]{mrm_tps_neighbourhood_recurrence_km()}}, \code{\link[rmorie:tpspn]{tpspn()}}}

  \item{rmoriebricklayer}{\code{\link[rmoriebricklayer:apply_schema_validation]{apply_schema_validation()}}, \code{\link[rmoriebricklayer:download_data]{download_data()}}, \code{\link[rmoriebricklayer:friendly_download]{friendly_download()}}, \code{\link[rmoriebricklayer:load_provenance]{load_provenance()}}, \code{\link[rmoriebricklayer:make_manifest]{make_manifest()}}, \code{\link[rmoriebricklayer:make_synthetic_column]{make_synthetic_column()}}, \code{\link[rmoriebricklayer:make_synthetic_csv]{make_synthetic_csv()}}, \code{\link[rmoriebricklayer:record]{record()}}, \code{\link[rmoriebricklayer:resolve_via_ckan]{resolve_via_ckan()}}, \code{\link[rmoriebricklayer:resolve_via_ckan_search]{resolve_via_ckan_search()}}, \code{\link[rmoriebricklayer:sha256_file]{sha256_file()}}, \code{\link[rmoriebricklayer:validate_schema]{validate_schema()}}, \code{\link[rmoriebricklayer:verify_sha256]{verify_sha256()}}, \code{\link[rmoriebricklayer:write_manifest_json]{write_manifest_json()}}, \code{\link[rmoriebricklayer:write_summary_txt]{write_summary_txt()}}}

  \item{rmoriedata}{\code{\link[rmoriedata:morie_data_checksums]{morie_data_checksums()}}}
}}

